SWIG refactor utils; clean up exported symbols; Update to swig git tip
