+++ python/setup.py
@@ -197,9 +197,11 @@ setup(
     platforms='Unix',
     py_modules=[
         'sentencepiece/__init__',
+        'sentencepiece/_utils',
         'sentencepiece/_version',
         'sentencepiece/sentencepiece_model_pb2',
         'sentencepiece/sentencepiece_pb2',
+        'sentencepiece/sentencepiece',
     ],
     ext_modules=[SENTENCEPIECE_EXT],
     cmdclass=cmdclass,
