+++ python/src/sentencepiece/sentencepiece_wrap.cxx
@@ -1,16 +1,17 @@
 /* ----------------------------------------------------------------------------
  * This file was automatically generated by SWIG (https://www.swig.org).
- * Version 4.1.0
+ * Version 4.4.0
  *
  * Do not make changes to this file unless you know what you are doing - modify
  * the SWIG interface file instead.
  * ----------------------------------------------------------------------------- */
 
 
-#define SWIG_VERSION 0x040100
+#define SWIG_VERSION 0x040400
 #define SWIGPYTHON
 #define SWIG_PYTHON_DIRECTOR_NO_VTABLE
 
+#define SWIG_name    "_sentencepiece"
 /* -----------------------------------------------------------------------------
  *  This section contains generic SWIG labels for method/variable
  *  declarations/attributes, and other compiler dependent labels.
@@ -135,6 +136,40 @@
 # pragma warning disable 592
 #endif
 
+#if defined(__cplusplus) && __cplusplus >=201103L
+# define SWIG_NOEXCEPT noexcept
+#else
+# define SWIG_NOEXCEPT throw()
+#endif 
+
+/* -----------------------------------------------------------------------------
+ * swigcompat.swg
+ *
+ * Macros to provide support compatibility with older C and C++ standards.
+ *
+ * Note that SWIG expects __cplusplus to be defined to the appropriate C++ standard.
+ * MSVC users are urged to check and examine the /Zc:__cplusplus compiler option.
+ * See https://learn.microsoft.com/en-us/cpp/build/reference/zc-cplusplus.
+ * ----------------------------------------------------------------------------- */
+
+/* C99 and C++11 should provide snprintf, but define SWIG_NO_SNPRINTF
+ * if you're missing it.
+ */
+#if ((defined __STDC_VERSION__ && __STDC_VERSION__ >= 199901L) || \
+     (defined __cplusplus && __cplusplus >= 201103L) || \
+     defined SWIG_HAVE_SNPRINTF) && \
+    !defined SWIG_NO_SNPRINTF
+# define SWIG_snprintf(O,S,F,A) snprintf(O,S,F,A)
+# define SWIG_snprintf2(O,S,F,A,B) snprintf(O,S,F,A,B)
+#else
+/* Fallback versions ignore the buffer size, but most of our uses either have a
+ * fixed maximum possible size or dynamically allocate a buffer that's large
+ * enough.
+ */
+# define SWIG_snprintf(O,S,F,A) sprintf(O,F,A)
+# define SWIG_snprintf2(O,S,F,A,B) sprintf(O,F,A,B)
+#endif
+
 
 #if defined(__GNUC__) && defined(_WIN32) && !defined(SWIG_PYTHON_NO_HYPOT_WORKAROUND)
 /* Workaround for '::hypot' has not been declared', see https://bugs.python.org/issue11566 */
@@ -168,9 +203,13 @@
 # include <Python.h>
 #endif
 
+
 #if __GNUC__ >= 7
 #pragma GCC diagnostic pop
 #endif
+
+#include <stdio.h>
+#include <stdlib.h>
 
 /* -----------------------------------------------------------------------------
  * swigrun.swg
@@ -181,7 +220,7 @@
 
 /* This should only be incremented when either the layout of swig_type_info changes,
    or for whatever reason, the runtime changes incompatibly */
-#define SWIG_RUNTIME_VERSION "4"
+#define SWIG_RUNTIME_VERSION "5"
 
 /* define SWIG_TYPE_TABLE_NAME as "SWIG_TYPE_TABLE" */
 #ifdef SWIG_TYPE_TABLE
@@ -319,7 +358,7 @@
 #define SWIG_CASTRANKLIMIT         (1 << 8)
 /* The NewMask denotes the object was created (using new/malloc) */
 #define SWIG_NEWOBJMASK            (SWIG_CASTRANKLIMIT  << 1)
-/* The TmpMask is for in/out typemaps that use temporal objects */
+/* The TmpMask is for in/out typemaps that use temporary objects */
 #define SWIG_TMPOBJMASK            (SWIG_NEWOBJMASK << 1)
 /* Simple returning values */
 #define SWIG_BADOBJ                (SWIG_ERROR)
@@ -379,8 +418,8 @@ typedef struct swig_type_info {
 typedef struct swig_cast_info {
   swig_type_info         *type;			/* pointer to type that is equivalent to this type */
   swig_converter_func     converter;		/* function to cast the void pointers */
-  struct swig_cast_info  *next;			/* pointer to next cast in linked list */
-  struct swig_cast_info  *prev;			/* pointer to the previous cast */
+  struct swig_cast_info  *next;			/* pointer to next array of casts | pointer to cast hashed by value */
+  unsigned int            value;		/* index of the last valid element in the array | typename hash value */
 } swig_cast_info;
 
 /* Structure used to store module information
@@ -442,54 +481,146 @@ SWIG_TypeEquiv(const char *nb, const char *tb) {
 }
 
 /*
+ * Hash function for type name strings, based on maRushPrime1Hash (http://amsoftware.narod.ru/algo2.html)
+ */
+SWIGRUNTIME unsigned int SWIG_Hash(const char *str, unsigned int len) {
+  const unsigned char *data = (const unsigned char *)str;
+  unsigned int hash = len, i = 0, k;
+  int rem = (int)len;
+
+  while (rem >= (int)sizeof(unsigned int)) {
+    k = *(unsigned int *)data;
+    k += i++;
+    hash ^= k;
+    hash *= 171717;
+    data += sizeof(unsigned int);
+    rem -= sizeof(unsigned int);
+  }
+
+  switch (rem) {
+    case 3: k = (unsigned int)(data[2]) << 16;
+            k |= (unsigned int)(data[1]) << 8;
+            k |= (unsigned int)(data[0]);
+            k += i++;
+            hash ^= k;
+            hash *= 171717;
+            break;
+    case 2: k = (unsigned int)(data[1]) << 8;
+            k |= (unsigned int)(data[0]);
+            k += i++;
+            hash ^= k;
+            hash *= 171717;
+            break;
+    case 1: k = (unsigned int)(data[0]);
+            k += i++;
+            hash ^= k;
+            hash *= 171717;
+            break;
+  }
+  return hash;
+}
+
+/*
   Check the typename
 */
 SWIGRUNTIME swig_cast_info *
 SWIG_TypeCheck(const char *c, swig_type_info *ty) {
+  static const unsigned int scan_threshold = 4;
   if (ty) {
-    swig_cast_info *iter = ty->cast;
-    while (iter) {
-      if (strcmp(iter->type->name, c) == 0) {
-        if (iter == ty->cast)
-          return iter;
-        /* Move iter to the top of the linked list */
-        iter->prev->next = iter->next;
-        if (iter->next)
-          iter->next->prev = iter->prev;
-        iter->next = ty->cast;
-        iter->prev = 0;
-        if (ty->cast) ty->cast->prev = iter;
-        ty->cast = iter;
-        return iter;
+    swig_cast_info *head = ty->cast;
+    unsigned int hash_value = 0;
+    int hashed = 0;
+
+    while (head) {
+
+      if (strcmp(head->type->name, c) == 0) {
+        return head;
       }
-      iter = iter->next;
+
+      if (head->value) {
+        swig_cast_info *iter;
+        swig_cast_info *last = head + head->value;
+        swig_cast_info *first = head + 1;
+        int search = 1;
+
+        if (!hashed) {
+          if (head->value < scan_threshold) {
+            for (iter = first; iter <= last; iter++) {
+              if (strcmp(iter->type->name, c) == 0) {
+                return iter;
+              }
+            }
+            search = 0;
+          } else {
+            hashed = 1;
+            hash_value = SWIG_Hash(c, (unsigned int)strlen(c));
+          }
+        }
+
+        if (search) {
+          /* Binary search over sorted <'next'|'value'> pairs */
+          do {
+            iter = first + ((last - first) >> 1);
+            if (iter->value < hash_value) {
+              first = iter + 1;
+            } else if (iter->value == hash_value) {
+
+              if (strcmp(iter->next->type->name, c) == 0) {
+                return iter->next;
+              }
+
+              /* Hash collision check */
+              for (last = iter + 1; last->next && last->value == hash_value; last++) {
+                if (strcmp(last->next->type->name, c) == 0) {
+                  return last->next;
+                }
+              }
+              for (first = iter - 1; first != head && first->value == hash_value; first--) {
+                if (strcmp(first->next->type->name, c) == 0) {
+                  return first->next;
+                }
+              }
+              break;
+            } else
+              last = iter - 1;
+          } while (first <= last);
+        }
+      }
+      head = head->next;
     }
   }
   return 0;
 }
 
 /*
-  Identical to SWIG_TypeCheck, except strcmp is replaced with a pointer comparison
+  Check the type by type address
 */
 SWIGRUNTIME swig_cast_info *
 SWIG_TypeCheckStruct(const swig_type_info *from, swig_type_info *ty) {
   if (ty) {
-    swig_cast_info *iter = ty->cast;
-    while (iter) {
-      if (iter->type == from) {
-        if (iter == ty->cast)
-          return iter;
-        /* Move iter to the top of the linked list */
-        iter->prev->next = iter->next;
-        if (iter->next)
-          iter->next->prev = iter->prev;
-        iter->next = ty->cast;
-        iter->prev = 0;
-        if (ty->cast) ty->cast->prev = iter;
-        ty->cast = iter;
-        return iter;
+    swig_cast_info *head = ty->cast;
+    while (head) {
+      if (head->type == from) {
+        return head;
       }
-      iter = iter->next;
+
+      if (head->value) {
+        swig_cast_info *iter;
+        swig_cast_info *last = head + head->value;
+        swig_cast_info *first = head + 1;
+
+        /* Binary search over sorted array of casts */
+        do {
+          iter = first + ((last - first) >> 1);
+          if (iter->type < from) {
+            first = iter + 1;
+          } else if (iter->type == from) {
+            return iter;
+          } else
+            last = iter - 1;
+        } while (first <= last);
+      }
+      head = head->next;
     }
   }
   return 0;
@@ -552,20 +683,24 @@ SWIG_TypePrettyName(const swig_type_info *type) {
 */
 SWIGRUNTIME void
 SWIG_TypeClientData(swig_type_info *ti, void *clientdata) {
-  swig_cast_info *cast = ti->cast;
+  swig_cast_info *head = ti->cast;
   /* if (ti->clientdata == clientdata) return; */
   ti->clientdata = clientdata;
 
-  while (cast) {
-    if (!cast->converter) {
-      swig_type_info *tc = cast->type;
-      if (!tc->clientdata) {
-	SWIG_TypeClientData(tc, clientdata);
+  while (head) {
+    swig_cast_info *cast;
+    for (cast = head; (cast - head) <= head->value; cast++) {
+      if (!cast->converter) {
+        swig_type_info *tc = cast->type;
+        if (!tc->clientdata) {
+          SWIG_TypeClientData(tc, clientdata);
+        }
       }
     }
-    cast = cast->next;
+    head = head->next;
   }
 }
+
 SWIGRUNTIME void
 SWIG_TypeNewClientData(swig_type_info *ti, void *clientdata) {
   SWIG_TypeClientData(ti, clientdata);
@@ -770,6 +905,12 @@ SWIG_UnpackDataName(const char *c, void *ptr, size_t sz, const char *name) {
 #define  SWIG_NullReferenceError   -13
 
 
+#if PY_VERSION_HEX >= 0x03030000 && !defined(SWIG_NO_HEAPTYPES)
+#if !defined(SWIG_HEAPTYPES)
+#define SWIG_HEAPTYPES
+#endif
+#endif
+
 /* Compatibility macros for Python 3 */
 #if PY_VERSION_HEX >= 0x03000000
 
@@ -789,10 +930,6 @@ SWIG_UnpackDataName(const char *c, void *ptr, size_t sz, const char *name) {
 
 #endif
 
-#ifndef Py_TYPE
-#  define Py_TYPE(op) ((op)->ob_type)
-#endif
-
 /* SWIG APIs for compatibility of both Python 2 & 3 */
 
 #if PY_VERSION_HEX >= 0x03000000
@@ -801,23 +938,42 @@ SWIG_UnpackDataName(const char *c, void *ptr, size_t sz, const char *name) {
 #  define SWIG_Python_str_FromFormat PyString_FromFormat
 #endif
 
+#if defined(SWIG_HEAPTYPES)
+#if PY_VERSION_HEX < 0x030c0000
+#include <structmember.h>
+#define Py_READONLY READONLY
+#define Py_T_PYSSIZET T_PYSSIZET
+#endif
+#endif
 
-SWIGINTERN char*
-SWIG_Python_str_AsChar(PyObject *str)
+#include <stddef.h> /* For offsetof */
+
+
+/* Wrapper around PyUnicode_AsUTF8AndSize - call Py_XDECREF on the returned pbytes when finished with the returned string */
+SWIGINTERN const char *
+SWIG_PyUnicode_AsUTF8AndSize(PyObject *str, Py_ssize_t *psize, PyObject **pbytes)
 {
 #if PY_VERSION_HEX >= 0x03030000
-  return (char *)PyUnicode_AsUTF8(str);
+# if !defined(Py_LIMITED_API) || Py_LIMITED_API+0 >= 0x030A0000
+  *pbytes = NULL;
+  return PyUnicode_AsUTF8AndSize(str, psize);
+# else
+  const char *chars;
+  *pbytes = PyUnicode_AsUTF8String(str);
+  chars = *pbytes ? PyBytes_AsString(*pbytes) : NULL;
+  if (chars && psize)
+    *psize = PyBytes_Size(*pbytes);
+  return chars;
+# endif
 #else
-  return PyString_AsString(str);
+  char *chars = NULL;
+  *pbytes = NULL;
+  PyString_AsStringAndSize(str, &chars, psize);
+  return chars;
 #endif
 }
 
-/* Was useful for Python 3.0.x-3.2.x - now provided only for compatibility
- * with any uses in user interface files. */
-#define SWIG_Python_str_DelForPy3(x)
-
-
-SWIGINTERN PyObject*
+SWIGINTERN PyObject *
 SWIG_Python_str_FromChar(const char *c)
 {
 #if PY_VERSION_HEX >= 0x03000000
@@ -827,9 +983,7 @@ SWIG_Python_str_FromChar(const char *c)
 #endif
 }
 
-#ifndef PyObject_DEL
-# define PyObject_DEL PyObject_Del
-#endif
+#define SWIG_RUNTIME_MODULE "swig_runtime_data" SWIG_RUNTIME_VERSION
 
 /* SWIGPY_USE_CAPSULE is no longer used within SWIG itself, but some user interface files check for it. */
 # define SWIGPY_USE_CAPSULE
@@ -838,12 +992,80 @@ SWIG_Python_str_FromChar(const char *c)
 #else
 # define SWIGPY_CAPSULE_ATTR_NAME "type_pointer_capsule" SWIG_TYPE_TABLE_NAME
 #endif
-# define SWIGPY_CAPSULE_NAME ("swig_runtime_data" SWIG_RUNTIME_VERSION "." SWIGPY_CAPSULE_ATTR_NAME)
+#define SWIGPY_CAPSULE_NAME SWIG_RUNTIME_MODULE "." SWIGPY_CAPSULE_ATTR_NAME
 
 #if PY_VERSION_HEX < 0x03020000
 #define PyDescr_TYPE(x) (((PyDescrObject *)(x))->d_type)
 #define PyDescr_NAME(x) (((PyDescrObject *)(x))->d_name)
 #define Py_hash_t long
+#endif
+
+#ifdef Py_GIL_DISABLED
+# undef PyTuple_GET_ITEM
+# undef PyTuple_SET_ITEM
+# undef PyTuple_GET_SIZE
+# undef PyCFunction_GET_FLAGS
+# undef PyCFunction_GET_FUNCTION
+# undef PyCFunction_GET_SELF
+# undef PyList_GET_ITEM
+# undef PyList_SET_ITEM
+# undef PySliceObject
+#endif
+
+
+#if defined(Py_LIMITED_API) || defined(Py_GIL_DISABLED)
+# define PyTuple_GET_ITEM PyTuple_GetItem
+/* Note that PyTuple_SetItem() has different semantics from PyTuple_SET_ITEM as it decref's the original tuple item, so in general they cannot be used
+  interchangeably. However in SWIG-generated code PyTuple_SET_ITEM is only used with newly initialized tuples without any items and for them this does work. */
+# define PyTuple_SET_ITEM PyTuple_SetItem
+# define PyTuple_GET_SIZE PyTuple_Size
+# define PyCFunction_GET_FLAGS PyCFunction_GetFlags
+# define PyCFunction_GET_FUNCTION PyCFunction_GetFunction
+# define PyCFunction_GET_SELF PyCFunction_GetSelf
+# define PyList_GET_ITEM PyList_GetItem
+# define PyList_SET_ITEM PyList_SetItem
+# define PySliceObject PyObject
+#endif
+
+/* Increment and Decrement wrappers - for portability when using the stable abi and for performance otherwise */
+#ifdef Py_LIMITED_API
+# define SWIG_Py_INCREF Py_IncRef
+# define SWIG_Py_XINCREF Py_IncRef
+# define SWIG_Py_DECREF Py_DecRef
+# define SWIG_Py_XDECREF Py_DecRef
+#else
+# define SWIG_Py_INCREF Py_INCREF
+# define SWIG_Py_XINCREF Py_XINCREF
+# define SWIG_Py_DECREF Py_DECREF
+# define SWIG_Py_XDECREF Py_XDECREF
+#endif
+
+#if PY_VERSION_HEX >= 0x03000000
+SWIGINTERN PyObject *
+SWIG_PyType_GetFullyQualifiedName(PyTypeObject *type) {
+#if PY_VERSION_HEX >= 0x030d0000
+  return PyType_GetFullyQualifiedName(type);
+#else
+  PyObject *result = NULL;
+  PyObject *qualname = PyObject_GetAttrString((PyObject *)type, "__qualname__");
+  if (qualname) {
+    PyObject *mod = PyObject_GetAttrString((PyObject *)type, "__module__");
+    if (mod) {
+      if (PyUnicode_Check(mod) && PyUnicode_CompareWithASCIIString(mod, "builtins") && PyUnicode_CompareWithASCIIString(mod, "__main__")) {
+        result = PyUnicode_FromFormat("%U%c%U", mod, '.', qualname);
+        SWIG_Py_DECREF(qualname);
+      } else {
+        result = qualname;
+      }
+      SWIG_Py_DECREF(mod);
+    } else {
+      result = qualname;
+    }
+  }
+
+  return result;
+#endif
+}
 #endif
 
 /* -----------------------------------------------------------------------------
@@ -887,6 +1109,9 @@ SWIG_Python_ErrorType(int code) {
   case SWIG_AttributeError:
     type = PyExc_AttributeError;
     break;
+  case SWIG_NullReferenceError:
+    type = PyExc_TypeError;
+    break;
   default:
     type = PyExc_RuntimeError;
   }
@@ -905,15 +1130,17 @@ SWIG_Python_AddErrorMsg(const char* mesg)
     PyErr_Fetch(&type, &value, &traceback);
   if (value) {
     PyObject *old_str = PyObject_Str(value);
-    const char *tmp = SWIG_Python_str_AsChar(old_str);
+    PyObject *bytes = NULL;
+    const char *tmp = SWIG_PyUnicode_AsUTF8AndSize(old_str, NULL, &bytes);
     PyErr_Clear();
-    Py_XINCREF(type);
+    SWIG_Py_XINCREF(type);
     if (tmp)
       PyErr_Format(type, "%s %s", tmp, mesg);
     else
       PyErr_Format(type, "%s", mesg);
-    Py_DECREF(old_str);
-    Py_DECREF(value);
+    SWIG_Py_XDECREF(bytes);
+    SWIG_Py_DECREF(old_str);
+    SWIG_Py_DECREF(value);
   } else {
     PyErr_SetString(PyExc_RuntimeError, mesg);
   }
@@ -943,7 +1170,7 @@ SWIG_Python_RaiseOrModifyTypeError(const char *message)
     newvalue = PyString_FromFormat("%s\nAdditional information:\n%s", PyString_AsString(value), message);
 #endif
     if (newvalue) {
-      Py_XDECREF(value);
+      SWIG_Py_XDECREF(value);
       PyErr_Restore(type, newvalue, traceback);
     } else {
       PyErr_Restore(type, value, traceback);
@@ -984,7 +1211,7 @@ SWIG_Python_RaiseOrModifyTypeError(const char *message)
          bool status;
          PyThreadState *save;
        public:
-         void end() { if (status) { PyEval_RestoreThread(save); status = false; }}
+         void end() { if (status) { status = false; PyEval_RestoreThread(save); }}
          SWIG_Python_Thread_Allow() : status(true), save(PyEval_SaveThread()) {}
          ~SWIG_Python_Thread_Allow() { end(); }
        };
@@ -1067,8 +1294,8 @@ typedef struct swig_const_info {
 # error "This version of SWIG only supports Python >= 2.7"
 #endif
 
-#if PY_VERSION_HEX >= 0x03000000 && PY_VERSION_HEX < 0x03030000
-# error "This version of SWIG only supports Python 3 >= 3.3"
+#if PY_VERSION_HEX >= 0x03000000 && PY_VERSION_HEX < 0x03050000
+# error "This version of SWIG only supports Python 3 >= 3.5"
 #endif
 
 /* Common SWIG API */
@@ -1119,7 +1346,6 @@ typedef struct swig_const_info {
 #define SWIG_Error(code, msg)            		SWIG_Python_SetErrorMsg(SWIG_ErrorType(code), msg) 
 #define SWIG_fail                        		goto fail					   
 
-
 /* Runtime API implementation */
 
 /* Error manipulation */
@@ -1128,7 +1354,7 @@ SWIGINTERN void
 SWIG_Python_SetErrorObj(PyObject *errtype, PyObject *obj) {
   SWIG_PYTHON_THREAD_BEGIN_BLOCK; 
   PyErr_SetObject(errtype, obj);
-  Py_DECREF(obj);
+  SWIG_Py_DECREF(obj);
   SWIG_PYTHON_THREAD_END_BLOCK;
 }
 
@@ -1149,13 +1375,13 @@ SWIGINTERN void
 SwigPyBuiltin_AddPublicSymbol(PyObject *seq, const char *key) {
   PyObject *s = PyString_InternFromString(key);
   PyList_Append(seq, s);
-  Py_DECREF(s);
+  SWIG_Py_DECREF(s);
 }
 
 SWIGINTERN void
 SWIG_Python_SetConstant(PyObject *d, PyObject *public_interface, const char *name, PyObject *obj) {   
   PyDict_SetItemString(d, name, obj);
-  Py_DECREF(obj);
+  SWIG_Py_DECREF(obj);
   if (public_interface)
     SwigPyBuiltin_AddPublicSymbol(public_interface, name);
 }
@@ -1165,19 +1391,36 @@ SWIG_Python_SetConstant(PyObject *d, PyObject *public_interface, const char *nam
 SWIGINTERN void
 SWIG_Python_SetConstant(PyObject *d, const char *name, PyObject *obj) {   
   PyDict_SetItemString(d, name, obj);
-  Py_DECREF(obj);                            
+  SWIG_Py_DECREF(obj);
 }
 
 #endif
 
-/* Append a value to the result obj */
+/* SWIG runtime data Python module */
+static PyObject *Swig_runtime_data_module_global = NULL;
 
+/* Create/obtain the single swig_runtime_data module which is used across different SWIG generated modules */
+SWIGINTERN PyObject *
+SWIG_runtime_data_module() {
+  if (!Swig_runtime_data_module_global) {
+#if PY_VERSION_HEX >= 0x03000000
+    Swig_runtime_data_module_global = PyImport_AddModule(SWIG_RUNTIME_MODULE);
+#else
+    static PyMethodDef swig_empty_runtime_method_table[] = { {NULL, NULL, 0, NULL} }; /* Sentinel */
+    Swig_runtime_data_module_global = Py_InitModule(SWIG_RUNTIME_MODULE, swig_empty_runtime_method_table);
+#endif
+    SWIG_Py_XINCREF(Swig_runtime_data_module_global);
+  }
+  return Swig_runtime_data_module_global;
+}
+
+/* Append a value to the result obj */
 SWIGINTERN PyObject*
-SWIG_Python_AppendOutput(PyObject* result, PyObject* obj) {
+SWIG_Python_AppendOutput(PyObject* result, PyObject* obj, int is_void) {
   if (!result) {
     result = obj;
-  } else if (result == Py_None) {
-    Py_DECREF(result);
+  } else if (result == Py_None && is_void) {
+    SWIG_Py_DECREF(result);
     result = obj;
   } else {
     if (!PyList_Check(result)) {
@@ -1186,12 +1429,12 @@ SWIG_Python_AppendOutput(PyObject* result, PyObject* obj) {
       if (result) {
         PyList_SET_ITEM(result, 0, o2);
       } else {
-        Py_DECREF(obj);
+        SWIG_Py_DECREF(obj);
         return o2;
       }
     }
     PyList_Append(result,obj);
-    Py_DECREF(obj);
+    SWIG_Py_DECREF(obj);
   }
   return result;
 }
@@ -1296,7 +1539,7 @@ typedef struct swig_varlinkobject {
 } swig_varlinkobject;
 
 SWIGINTERN PyObject *
-swig_varlink_repr(PyObject *SWIGUNUSEDPARM(v)) {
+SwigVarLink_repr(PyObject *SWIGUNUSEDPARM(v)) {
 #if PY_VERSION_HEX >= 0x03000000
   return PyUnicode_InternFromString("<Swig global variables>");
 #else
@@ -1305,7 +1548,7 @@ swig_varlink_repr(PyObject *SWIGUNUSEDPARM(v)) {
 }
 
 SWIGINTERN PyObject *
-swig_varlink_str(PyObject *o) {
+SwigVarLink_str(PyObject *o) {
   swig_varlinkobject *v = (swig_varlinkobject *) o;
 #if PY_VERSION_HEX >= 0x03000000
   PyObject *str = PyUnicode_InternFromString("(");
@@ -1315,21 +1558,21 @@ swig_varlink_str(PyObject *o) {
   for (var = v->vars; var; var=var->next) {
     tail = PyUnicode_FromString(var->name);
     joined = PyUnicode_Concat(str, tail);
-    Py_DecRef(str);
-    Py_DecRef(tail);
+    SWIG_Py_DECREF(str);
+    SWIG_Py_DECREF(tail);
     str = joined;
     if (var->next) {
         tail = PyUnicode_InternFromString(", ");
         joined = PyUnicode_Concat(str, tail);
-        Py_DecRef(str);
-        Py_DecRef(tail);
+        SWIG_Py_DECREF(str);
+        SWIG_Py_DECREF(tail);
         str = joined;
     }
   }
   tail = PyUnicode_InternFromString(")");
   joined = PyUnicode_Concat(str, tail);
-  Py_DecRef(str);
-  Py_DecRef(tail);
+  SWIG_Py_DECREF(str);
+  SWIG_Py_DECREF(tail);
   str = joined;
 #else
   PyObject *str = PyString_FromString("(");
@@ -1344,7 +1587,7 @@ swig_varlink_str(PyObject *o) {
 }
 
 SWIGINTERN void
-swig_varlink_dealloc(PyObject *o) {
+SwigVarLink_dealloc(PyObject *o) {
   swig_varlinkobject *v = (swig_varlinkobject *) o;
   swig_globalvar *var = v->vars;
   while (var) {
@@ -1356,7 +1599,7 @@ swig_varlink_dealloc(PyObject *o) {
 }
 
 SWIGINTERN PyObject *
-swig_varlink_getattr(PyObject *o, char *n) {
+SwigVarLink_getattr(PyObject *o, char *n) {
   swig_varlinkobject *v = (swig_varlinkobject *) o;
   PyObject *res = NULL;
   swig_globalvar *var = v->vars;
@@ -1374,7 +1617,7 @@ swig_varlink_getattr(PyObject *o, char *n) {
 }
 
 SWIGINTERN int
-swig_varlink_setattr(PyObject *o, char *n, PyObject *p) {
+SwigVarLink_setattr(PyObject *o, char *n, PyObject *p) {
   swig_varlinkobject *v = (swig_varlinkobject *) o;
   int res = 1;
   swig_globalvar *var = v->vars;
@@ -1392,8 +1635,9 @@ swig_varlink_setattr(PyObject *o, char *n, PyObject *p) {
 }
 
 SWIGINTERN PyTypeObject*
-swig_varlink_type(void) {
-  static char varlink__doc__[] = "Swig var link object";
+SwigVarLink_TypeOnce(void) {
+  static char SwigVarLink_doc[] = "Swig variable link object";
+#ifndef SWIG_HEAPTYPES
   static PyTypeObject varlink_type;
   static int type_init = 0;
   if (!type_init) {
@@ -1404,30 +1648,30 @@ swig_varlink_type(void) {
       PyObject_HEAD_INIT(NULL)
       0,                                  /* ob_size */
 #endif
-      "swigvarlink",                      /* tp_name */
+      SWIG_RUNTIME_MODULE ".SwigVarLink", /* tp_name */
       sizeof(swig_varlinkobject),         /* tp_basicsize */
       0,                                  /* tp_itemsize */
-      (destructor) swig_varlink_dealloc,  /* tp_dealloc */
+      (destructor) SwigVarLink_dealloc,   /* tp_dealloc */
 #if PY_VERSION_HEX < 0x030800b4
-      (printfunc)0,                       /*tp_print*/
+      (printfunc)0,                       /* tp_print */
 #else
-      (Py_ssize_t)0,                      /*tp_vectorcall_offset*/
+      (Py_ssize_t)0,                      /* tp_vectorcall_offset */
 #endif
-      (getattrfunc) swig_varlink_getattr, /* tp_getattr */
-      (setattrfunc) swig_varlink_setattr, /* tp_setattr */
+      (getattrfunc) SwigVarLink_getattr,  /* tp_getattr */
+      (setattrfunc) SwigVarLink_setattr,  /* tp_setattr */
       0,                                  /* tp_compare */
-      (reprfunc) swig_varlink_repr,       /* tp_repr */
+      (reprfunc) SwigVarLink_repr,        /* tp_repr */
       0,                                  /* tp_as_number */
       0,                                  /* tp_as_sequence */
       0,                                  /* tp_as_mapping */
       0,                                  /* tp_hash */
       0,                                  /* tp_call */
-      (reprfunc) swig_varlink_str,        /* tp_str */
+      (reprfunc) SwigVarLink_str,         /* tp_str */
       0,                                  /* tp_getattro */
       0,                                  /* tp_setattro */
       0,                                  /* tp_as_buffer */
       0,                                  /* tp_flags */
-      varlink__doc__,                     /* tp_doc */
+      SwigVarLink_doc,                    /* tp_doc */
       0,                                  /* tp_traverse */
       0,                                  /* tp_clear */
       0,                                  /* tp_richcompare */
@@ -1444,6 +1688,12 @@ swig_varlink_type(void) {
 #if (PY_VERSION_HEX >= 0x03080000) && (PY_VERSION_HEX < 0x03090000)
       0,                                  /* tp_print */
 #endif
+#if PY_VERSION_HEX >= 0x030c0000
+      0,                                  /* tp_watched */
+#endif
+#if PY_VERSION_HEX >= 0x030d00a4
+      0,                                  /* tp_versions_used */
+#endif
 #ifdef COUNT_ALLOCS
       0,                                  /* tp_allocs */
       0,                                  /* tp_frees */
@@ -1452,18 +1702,50 @@ swig_varlink_type(void) {
       0                                   /* tp_next */
 #endif
     };
+    PyObject *runtime_data_module = SWIG_runtime_data_module();
     varlink_type = tmp;
     type_init = 1;
     if (PyType_Ready(&varlink_type) < 0)
       return NULL;
+    if (PyModule_AddObject(runtime_data_module, "SwigVarLink", (PyObject *)&varlink_type) == 0)
+      SWIG_Py_INCREF((PyObject *)&varlink_type);
   }
   return &varlink_type;
+#else
+  PyType_Slot slots[] = {
+    { Py_tp_dealloc, (void *)SwigVarLink_dealloc },
+    { Py_tp_repr, (void *)SwigVarLink_repr },
+    { Py_tp_getattr, (void *)SwigVarLink_getattr },
+    { Py_tp_setattr, (void *)SwigVarLink_setattr },
+    { Py_tp_str, (void *)SwigVarLink_str },
+    { Py_tp_doc, (void *)SwigVarLink_doc },
+    { 0, NULL }
+  };
+  PyType_Spec spec = {
+    SWIG_RUNTIME_MODULE ".SwigVarLink",
+    sizeof(swig_varlinkobject),
+    0,
+    Py_TPFLAGS_DEFAULT,
+    slots
+  };
+  PyObject *pytype = PyType_FromSpec(&spec);
+  PyObject *runtime_data_module = SWIG_runtime_data_module();
+  if (pytype && PyModule_AddObject(runtime_data_module, "SwigVarLink", pytype) == 0)
+    SWIG_Py_INCREF(pytype);
+  return (PyTypeObject *)pytype;
+#endif
+}
+
+SWIGRUNTIME PyTypeObject*
+SwigVarLink_Type(void) {
+  static PyTypeObject *SWIG_STATIC_POINTER(type) = SwigVarLink_TypeOnce();
+  return type;
 }
 
 /* Create a variable linking object for use later */
 SWIGINTERN PyObject *
 SWIG_Python_newvarlink(void) {
-  swig_varlinkobject *result = PyObject_NEW(swig_varlinkobject, swig_varlink_type());
+  swig_varlinkobject *result = PyObject_New(swig_varlinkobject, SwigVarLink_Type());
   if (result) {
     result->vars = 0;
   }
@@ -1525,7 +1807,7 @@ SWIGRUNTIMEINLINE PyObject *
 SWIG_Py_Void(void)
 {
   PyObject *none = Py_None;
-  Py_INCREF(none);
+  SWIG_Py_INCREF(none);
   return none;
 }
 
@@ -1568,27 +1850,27 @@ SwigPyClientData_New(PyObject* obj)
     SwigPyClientData *data = (SwigPyClientData *)malloc(sizeof(SwigPyClientData));
     /* the klass element */
     data->klass = obj;
-    Py_INCREF(data->klass);
+    SWIG_Py_INCREF(data->klass);
     /* the newraw method and newargs arguments used to create a new raw instance */
     if (PyClass_Check(obj)) {
       data->newraw = 0;
-      Py_INCREF(obj);
+      SWIG_Py_INCREF(obj);
       data->newargs = obj;
     } else {
       data->newraw = PyObject_GetAttrString(data->klass, "__new__");
       if (data->newraw) {
         data->newargs = PyTuple_New(1);
         if (data->newargs) {
-          Py_INCREF(obj);
+          SWIG_Py_INCREF(obj);
           PyTuple_SET_ITEM(data->newargs, 0, obj);
         } else {
-          Py_DECREF(data->newraw);
-          Py_DECREF(data->klass);
+          SWIG_Py_DECREF(data->newraw);
+          SWIG_Py_DECREF(data->klass);
           free(data);
           return 0;
         }
       } else {
-        Py_INCREF(obj);
+        SWIG_Py_INCREF(obj);
         data->newargs = obj;
       }
     }
@@ -1612,10 +1894,10 @@ SwigPyClientData_New(PyObject* obj)
 SWIGRUNTIME void 
 SwigPyClientData_Del(SwigPyClientData *data)
 {
-  Py_XDECREF(data->klass);
-  Py_XDECREF(data->newraw);
-  Py_XDECREF(data->newargs);
-  Py_XDECREF(data->destroy);
+  SWIG_Py_XDECREF(data->klass);
+  SWIG_Py_XDECREF(data->newraw);
+  SWIG_Py_XDECREF(data->newargs);
+  SWIG_Py_XDECREF(data->destroy);
   free(data);
 }
 
@@ -1627,9 +1909,7 @@ typedef struct {
   swig_type_info *ty;
   int own;
   PyObject *next;
-#ifdef SWIGPYTHON_BUILTIN
-  PyObject *dict;
-#endif
+  PyObject *swigdict;
 } SwigPyObject;
 
 
@@ -1640,11 +1920,11 @@ SwigPyObject_get___dict__(PyObject *v, PyObject *SWIGUNUSEDPARM(args))
 {
   SwigPyObject *sobj = (SwigPyObject *)v;
 
-  if (!sobj->dict)
-    sobj->dict = PyDict_New();
+  if (!sobj->swigdict)
+    sobj->swigdict = PyDict_New();
 
-  Py_XINCREF(sobj->dict);
-  return sobj->dict;
+  SWIG_Py_XINCREF(sobj->swigdict);
+  return sobj->swigdict;
 }
 
 #endif
@@ -1672,10 +1952,10 @@ SwigPyObject_format(const char* fmt, SwigPyObject *v)
 #else
         res = PyString_Format(ofmt,args);
 #endif
-        Py_DECREF(ofmt);
+        SWIG_Py_DECREF(ofmt);
       }
     }
-    Py_DECREF(args);
+    SWIG_Py_DECREF(args);
   }
   return res;
 }
@@ -1702,14 +1982,14 @@ SwigPyObject_repr(SwigPyObject *v)
     if (nrep) {
 # if PY_VERSION_HEX >= 0x03000000
       PyObject *joined = PyUnicode_Concat(repr, nrep);
-      Py_DecRef(repr);
-      Py_DecRef(nrep);
+      SWIG_Py_DECREF(repr);
+      SWIG_Py_DECREF(nrep);
       repr = joined;
 # else
       PyString_ConcatAndDel(&repr,nrep);
 # endif
     } else {
-      Py_DecRef(repr);
+      SWIG_Py_DECREF(repr);
       repr = NULL;
     }
   }
@@ -1717,7 +1997,7 @@ SwigPyObject_repr(SwigPyObject *v)
 }
 
 /* We need a version taking two PyObject* parameters so it's a valid
- * PyCFunction to use in swigobject_methods[]. */
+ * PyCFunction to use in SwigPyObject_methods[]. */
 SWIGRUNTIME PyObject *
 SwigPyObject_repr2(PyObject *v, PyObject *SWIGUNUSEDPARM(args))
 {
@@ -1725,24 +2005,33 @@ SwigPyObject_repr2(PyObject *v, PyObject *SWIGUNUSEDPARM(args))
 }
 
 SWIGRUNTIME int
-SwigPyObject_compare(SwigPyObject *v, SwigPyObject *w)
+SwigPyObject_compare(PyObject *v, PyObject *w)
 {
-  void *i = v->ptr;
-  void *j = w->ptr;
+  /* tp_compare is only called when both objects have the same type, so
+   * the casts are guaranteed to be ok. */
+  void *i = ((SwigPyObject *)v)->ptr;
+  void *j = ((SwigPyObject *)w)->ptr;
   return (i < j) ? -1 : ((i > j) ? 1 : 0);
 }
 
+SWIGRUNTIMEINLINE int SwigPyObject_Check(PyObject *);
+
 /* Added for Python 3.x, would it also be useful for Python 2.x? */
 SWIGRUNTIME PyObject*
-SwigPyObject_richcompare(SwigPyObject *v, SwigPyObject *w, int op)
+SwigPyObject_richcompare(PyObject *v, PyObject *w, int op)
 {
-  PyObject* res;
-  if( op != Py_EQ && op != Py_NE ) {
-    Py_INCREF(Py_NotImplemented);
-    return Py_NotImplemented;
+  PyObject* res = NULL;
+  if (!PyErr_Occurred()) {
+    /* Per https://docs.python.org/3/c-api/typeobj.html#c.PyTypeObject.tp_richcompare
+     * the first argument is guaranteed to be an instance of SwigPyObject, but the
+     * second is not, so we typecheck that one. */
+    if ((op != Py_EQ && op != Py_NE) || !SwigPyObject_Check(w)) {
+      SWIG_Py_INCREF(Py_NotImplemented);
+      return Py_NotImplemented;
+    }
+    res = PyBool_FromLong( (SwigPyObject_compare(v, w)==0) == (op == Py_EQ) ? 1 : 0);
   }
-  res = PyBool_FromLong( (SwigPyObject_compare(v, w)==0) == (op == Py_EQ) ? 1 : 0);
-  return res;  
+  return res;
 }
 
 
@@ -1751,7 +2040,7 @@ SWIGRUNTIME PyTypeObject* SwigPyObject_TypeOnce(void);
 #ifdef SWIGPYTHON_BUILTIN
 static swig_type_info *SwigPyObject_stype = 0;
 SWIGRUNTIME PyTypeObject*
-SwigPyObject_type(void) {
+SwigPyObject_Type(void) {
     SwigPyClientData *cd;
     assert(SwigPyObject_stype);
     cd = (SwigPyClientData*) SwigPyObject_stype->clientdata;
@@ -1761,7 +2050,7 @@ SwigPyObject_type(void) {
 }
 #else
 SWIGRUNTIME PyTypeObject*
-SwigPyObject_type(void) {
+SwigPyObject_Type(void) {
   static PyTypeObject *SWIG_STATIC_POINTER(type) = SwigPyObject_TypeOnce();
   return type;
 }
@@ -1769,14 +2058,29 @@ SwigPyObject_type(void) {
 
 SWIGRUNTIMEINLINE int
 SwigPyObject_Check(PyObject *op) {
+  PyTypeObject *target_tp = SwigPyObject_Type();
+  PyTypeObject *op_type = Py_TYPE(op);
 #ifdef SWIGPYTHON_BUILTIN
-  PyTypeObject *target_tp = SwigPyObject_type();
-  if (PyType_IsSubtype(op->ob_type, target_tp))
-    return 1;
-  return (strcmp(op->ob_type->tp_name, "SwigPyObject") == 0);
+  /* Only builtin types have SwigPyObject as a base type */
+  return PyType_IsSubtype(op_type, target_tp);
 #else
-  return (Py_TYPE(op) == SwigPyObject_type())
-    || (strcmp(Py_TYPE(op)->tp_name,"SwigPyObject") == 0);
+  /* Check for an exact match to SwigPyObject */
+  if (op_type == target_tp) {
+    return 1;
+  } else {
+    /* Fallback for multiple modules */
+#if PY_VERSION_HEX >= 0x03000000
+    int cmp;
+    PyObject *tpname = SWIG_PyType_GetFullyQualifiedName(op_type);
+    if (!tpname)
+      return 0;
+    cmp = PyUnicode_CompareWithASCIIString(tpname, SWIG_RUNTIME_MODULE ".SwigPyObject");
+    SWIG_Py_DECREF(tpname);
+    return cmp == 0;
+#else
+    return strcmp(op_type->tp_name, SWIG_RUNTIME_MODULE ".SwigPyObject") == 0;
+#endif
+  }
 #endif
 }
 
@@ -1816,7 +2120,7 @@ SwigPyObject_dealloc(PyObject *v)
         } else {
           res = 0;
         }
-        Py_XDECREF(tmp);
+        SWIG_Py_XDECREF(tmp);
       } else {
         PyCFunction meth = PyCFunction_GET_FUNCTION(destroy);
         PyObject *mself = PyCFunction_GET_SELF(destroy);
@@ -1827,7 +2131,7 @@ SwigPyObject_dealloc(PyObject *v)
 
       PyErr_Restore(type, value, traceback);
 
-      Py_XDECREF(res);
+      SWIG_Py_XDECREF(res);
     } 
 #if !defined(SWIG_PYTHON_SILENT_MEMLEAK)
     else {
@@ -1835,13 +2139,11 @@ SwigPyObject_dealloc(PyObject *v)
       printf("swig/python detected a memory leak of type '%s', no destructor found.\n", (name ? name : "unknown"));
     }
 #endif
-    Py_XDECREF(Swig_Capsule_global);
+    SWIG_Py_XDECREF(Swig_Capsule_global);
   }
-  Py_XDECREF(next);
-#ifdef SWIGPYTHON_BUILTIN
-  Py_XDECREF(sobj->dict);
-#endif
-  PyObject_DEL(v);
+  SWIG_Py_XDECREF(next);
+  SWIG_Py_XDECREF(sobj->swigdict);
+  PyObject_Free(v);
 }
 
 SWIGRUNTIME PyObject* 
@@ -1854,7 +2156,7 @@ SwigPyObject_append(PyObject* v, PyObject* next)
   }
   ((SwigPyObject *)next)->next = sobj->next;
   sobj->next = next;
-  Py_INCREF(next);
+  SWIG_Py_INCREF(next);
   return SWIG_Py_Void();
 }
 
@@ -1863,7 +2165,7 @@ SwigPyObject_next(PyObject* v, PyObject *SWIGUNUSEDPARM(args))
 {
   SwigPyObject *sobj = (SwigPyObject *) v;
   if (sobj->next) {    
-    Py_INCREF(sobj->next);
+    SWIG_Py_INCREF(sobj->next);
     return sobj->next;
   } else {
     return SWIG_Py_Void();
@@ -1897,9 +2199,9 @@ SwigPyObject_own(PyObject *v, PyObject *args)
     PyObject *obj = PyBool_FromLong(sobj->own);
     if (val) {
       if (PyObject_IsTrue(val)) {
-        Py_DECREF(SwigPyObject_acquire(v,args));
+        SWIG_Py_DECREF(SwigPyObject_acquire(v,args));
       } else {
-        Py_DECREF(SwigPyObject_disown(v,args));
+        SWIG_Py_DECREF(SwigPyObject_disown(v,args));
       }
     } 
     return obj;
@@ -1907,7 +2209,7 @@ SwigPyObject_own(PyObject *v, PyObject *args)
 }
 
 static PyMethodDef
-swigobject_methods[] = {
+SwigPyObject_methods[] = {
   {"disown",  SwigPyObject_disown,  METH_NOARGS,  "releases ownership of the pointer"},
   {"acquire", SwigPyObject_acquire, METH_NOARGS,  "acquires ownership of the pointer"},
   {"own",     SwigPyObject_own,     METH_VARARGS, "returns/sets ownership of the pointer"},
@@ -1919,8 +2221,8 @@ swigobject_methods[] = {
 
 SWIGRUNTIME PyTypeObject*
 SwigPyObject_TypeOnce(void) {
-  static char swigobject_doc[] = "Swig object carries a C/C++ instance pointer";
-
+  static char SwigPyObject_doc[] = "Swig object holding a C/C++ pointer";
+#ifndef SWIG_HEAPTYPES
   static PyNumberMethods SwigPyObject_as_number = {
     (binaryfunc)0, /*nb_add*/
     (binaryfunc)0, /*nb_subtract*/
@@ -1975,14 +2277,14 @@ SwigPyObject_TypeOnce(void) {
       PyObject_HEAD_INIT(NULL)
       0,                                    /* ob_size */
 #endif
-      "SwigPyObject",                       /* tp_name */
+      SWIG_RUNTIME_MODULE ".SwigPyObject",  /* tp_name */
       sizeof(SwigPyObject),                 /* tp_basicsize */
       0,                                    /* tp_itemsize */
       (destructor)SwigPyObject_dealloc,     /* tp_dealloc */
 #if PY_VERSION_HEX < 0x030800b4
-      (printfunc)0,                         /*tp_print*/
+      (printfunc)0,                         /* tp_print */
 #else
-      (Py_ssize_t)0,                        /*tp_vectorcall_offset*/
+      (Py_ssize_t)0,                        /* tp_vectorcall_offset */
 #endif
       (getattrfunc)0,                       /* tp_getattr */
       (setattrfunc)0,                       /* tp_setattr */
@@ -2002,14 +2304,14 @@ SwigPyObject_TypeOnce(void) {
       0,                                    /* tp_setattro */
       0,                                    /* tp_as_buffer */
       Py_TPFLAGS_DEFAULT,                   /* tp_flags */
-      swigobject_doc,                       /* tp_doc */
+      SwigPyObject_doc,                     /* tp_doc */
       0,                                    /* tp_traverse */
       0,                                    /* tp_clear */
       (richcmpfunc)SwigPyObject_richcompare,/* tp_richcompare */
       0,                                    /* tp_weaklistoffset */
       0,                                    /* tp_iter */
       0,                                    /* tp_iternext */
-      swigobject_methods,                   /* tp_methods */
+      SwigPyObject_methods,                 /* tp_methods */
       0,                                    /* tp_members */
       0,                                    /* tp_getset */
       0,                                    /* tp_base */
@@ -2038,6 +2340,12 @@ SwigPyObject_TypeOnce(void) {
 #if (PY_VERSION_HEX >= 0x03080000) && (PY_VERSION_HEX < 0x03090000)
       0,                                    /* tp_print */
 #endif
+#if PY_VERSION_HEX >= 0x030c0000
+      0,                                    /* tp_watched */
+#endif
+#if PY_VERSION_HEX >= 0x030d00a4
+      0,                                    /* tp_versions_used */
+#endif
 #ifdef COUNT_ALLOCS
       0,                                    /* tp_allocs */
       0,                                    /* tp_frees */
@@ -2046,31 +2354,70 @@ SwigPyObject_TypeOnce(void) {
       0                                     /* tp_next */
 #endif
     };
+    PyObject *runtime_data_module = SWIG_runtime_data_module();
     swigpyobject_type = tmp;
     type_init = 1;
     if (PyType_Ready(&swigpyobject_type) != 0)
       return NULL;
+    if (PyModule_AddObject(runtime_data_module, "SwigPyObject", (PyObject *)&swigpyobject_type) == 0)
+      SWIG_Py_INCREF((PyObject *)&swigpyobject_type);
   }
   return &swigpyobject_type;
+#else
+  static PyMemberDef SwigPyObject_members[] = {
+      { (char *)"__dictoffset__", Py_T_PYSSIZET, offsetof(SwigPyObject, swigdict), Py_READONLY, NULL },
+      { NULL, 0, 0, 0, NULL }
+  };
+  PyType_Slot slots[] = {
+    { Py_tp_dealloc, (void *)SwigPyObject_dealloc },
+    { Py_tp_repr, (void *)SwigPyObject_repr },
+    { Py_tp_getattro, (void *)PyObject_GenericGetAttr },
+    { Py_tp_doc, (void *)SwigPyObject_doc },
+    { Py_tp_richcompare, (void *)SwigPyObject_richcompare },
+    { Py_tp_methods, (void *)SwigPyObject_methods },
+    { Py_nb_int, (void *)SwigPyObject_long },
+    { Py_tp_members, (void *)SwigPyObject_members },
+    { 0, NULL }
+  };
+  PyType_Spec spec = {
+    SWIG_RUNTIME_MODULE ".SwigPyObject",
+    sizeof(SwigPyObject),
+    0,
+    Py_TPFLAGS_DEFAULT|Py_TPFLAGS_BASETYPE,
+    slots
+  };
+  PyObject *pytype = PyType_FromSpec(&spec);
+  PyObject *runtime_data_module = SWIG_runtime_data_module();
+#if !defined(Py_LIMITED_API)
+/* While this __dictoffset__ is only used with the builtin wrappers, SwigPyObject ought to be
+   identical when created for use by proxy class wrappers in case it is shared across multiple modules. */
+#if PY_VERSION_HEX < 0x03090000
+    /* Workaround as __dictoffset__ above is only supported from python-3.9 */
+  if (pytype)
+    ((PyTypeObject *)pytype)->tp_dictoffset = offsetof(SwigPyObject, swigdict);
+#endif
+#endif
+  if (pytype && PyModule_AddObject(runtime_data_module, "SwigPyObject", pytype) == 0)
+    SWIG_Py_INCREF(pytype);
+  return (PyTypeObject *)pytype;
+#endif
 }
 
 SWIGRUNTIME PyObject *
 SwigPyObject_New(void *ptr, swig_type_info *ty, int own)
 {
-  SwigPyObject *sobj = PyObject_NEW(SwigPyObject, SwigPyObject_type());
+  SwigPyObject *sobj = PyObject_New(SwigPyObject, SwigPyObject_Type());
   if (sobj) {
     sobj->ptr  = ptr;
     sobj->ty   = ty;
     sobj->own  = own;
     sobj->next = 0;
-#ifdef SWIGPYTHON_BUILTIN
-    sobj->dict = 0;
-#endif
+    sobj->swigdict = 0;
     if (own == SWIG_POINTER_OWN) {
       /* Obtain a reference to the Python capsule wrapping the module information, so that the
        * module information is correctly destroyed after all SWIG python objects have been freed
        * by the GC (and corresponding destructors invoked) */
-      Py_XINCREF(Swig_Capsule_global);
+      SWIG_Py_XINCREF(Swig_Capsule_global);
     }
   }
   return (PyObject *)sobj;
@@ -2121,15 +2468,32 @@ SwigPyPacked_compare(SwigPyPacked *v, SwigPyPacked *w)
 SWIGRUNTIME PyTypeObject* SwigPyPacked_TypeOnce(void);
 
 SWIGRUNTIME PyTypeObject*
-SwigPyPacked_type(void) {
+SwigPyPacked_Type(void) {
   static PyTypeObject *SWIG_STATIC_POINTER(type) = SwigPyPacked_TypeOnce();
   return type;
 }
 
 SWIGRUNTIMEINLINE int
 SwigPyPacked_Check(PyObject *op) {
-  return ((op)->ob_type == SwigPyPacked_TypeOnce()) 
-    || (strcmp((op)->ob_type->tp_name,"SwigPyPacked") == 0);
+  PyTypeObject *target_tp = SwigPyPacked_Type();
+  PyTypeObject *op_type = Py_TYPE(op);
+  /* Check for an exact match to SwigPyPacked */
+  if (op_type == target_tp) {
+    return 1;
+  } else {
+    /* Fallback for multiple modules */
+#if PY_VERSION_HEX >= 0x03000000
+    int cmp;
+    PyObject *tpname = SWIG_PyType_GetFullyQualifiedName(op_type);
+    if (!tpname)
+      return 0;
+    cmp = PyUnicode_CompareWithASCIIString(tpname, SWIG_RUNTIME_MODULE ".SwigPyPacked");
+    SWIG_Py_DECREF(tpname);
+    return cmp == 0;
+#else
+    return strcmp(op_type->tp_name, SWIG_RUNTIME_MODULE ".SwigPyPacked") == 0;
+#endif
+  }
 }
 
 SWIGRUNTIME void
@@ -2139,34 +2503,35 @@ SwigPyPacked_dealloc(PyObject *v)
     SwigPyPacked *sobj = (SwigPyPacked *) v;
     free(sobj->pack);
   }
-  PyObject_DEL(v);
+  PyObject_Free(v);
 }
 
 SWIGRUNTIME PyTypeObject*
 SwigPyPacked_TypeOnce(void) {
-  static char swigpacked_doc[] = "Swig object carries a C/C++ instance pointer";
+  static char SwigPyPacked_doc[] = "Swig object holding a C/C++ function pointer";
+#ifndef SWIG_HEAPTYPES
   static PyTypeObject swigpypacked_type;
   static int type_init = 0;
   if (!type_init) {
     const PyTypeObject tmp = {
-#if PY_VERSION_HEX>=0x03000000
+#if PY_VERSION_HEX >= 0x03000000
       PyVarObject_HEAD_INIT(NULL, 0)
 #else
       PyObject_HEAD_INIT(NULL)
       0,                                    /* ob_size */
 #endif
-      "SwigPyPacked",                       /* tp_name */
+      SWIG_RUNTIME_MODULE ".SwigPyPacked",  /* tp_name */
       sizeof(SwigPyPacked),                 /* tp_basicsize */
       0,                                    /* tp_itemsize */
       (destructor)SwigPyPacked_dealloc,     /* tp_dealloc */
 #if PY_VERSION_HEX < 0x030800b4
-      (printfunc)0,                         /*tp_print*/
+      (printfunc)0,                         /* tp_print */
 #else
-      (Py_ssize_t)0,                        /*tp_vectorcall_offset*/
+      (Py_ssize_t)0,                        /* tp_vectorcall_offset */
 #endif
       (getattrfunc)0,                       /* tp_getattr */
       (setattrfunc)0,                       /* tp_setattr */
-#if PY_VERSION_HEX>=0x03000000
+#if PY_VERSION_HEX >= 0x03000000
       0, /* tp_reserved in 3.0.1 */
 #else
       (cmpfunc)SwigPyPacked_compare,        /* tp_compare */
@@ -2182,7 +2547,7 @@ SwigPyPacked_TypeOnce(void) {
       0,                                    /* tp_setattro */
       0,                                    /* tp_as_buffer */
       Py_TPFLAGS_DEFAULT,                   /* tp_flags */
-      swigpacked_doc,                       /* tp_doc */
+      SwigPyPacked_doc,                     /* tp_doc */
       0,                                    /* tp_traverse */
       0,                                    /* tp_clear */
       0,                                    /* tp_richcompare */
@@ -2218,6 +2583,12 @@ SwigPyPacked_TypeOnce(void) {
 #if (PY_VERSION_HEX >= 0x03080000) && (PY_VERSION_HEX < 0x03090000)
       0,                                    /* tp_print */
 #endif
+#if PY_VERSION_HEX >= 0x030c0000
+      0,                                    /* tp_watched */
+#endif
+#if PY_VERSION_HEX >= 0x030d00a4
+      0,                                    /* tp_versions_used */
+#endif
 #ifdef COUNT_ALLOCS
       0,                                    /* tp_allocs */
       0,                                    /* tp_frees */
@@ -2226,18 +2597,43 @@ SwigPyPacked_TypeOnce(void) {
       0                                     /* tp_next */
 #endif
     };
+    PyObject *runtime_data_module = SWIG_runtime_data_module();
     swigpypacked_type = tmp;
     type_init = 1;
     if (PyType_Ready(&swigpypacked_type) != 0)
       return NULL;
+    if (PyModule_AddObject(runtime_data_module, "SwigPyPacked", (PyObject *)&swigpypacked_type) == 0)
+      SWIG_Py_INCREF((PyObject *)&swigpypacked_type);
   }
   return &swigpypacked_type;
+#else
+  PyType_Slot slots[] = {
+    { Py_tp_dealloc, (void *)SwigPyPacked_dealloc },
+    { Py_tp_repr, (void *)SwigPyPacked_repr },
+    { Py_tp_str, (void *)SwigPyPacked_str },
+    { Py_tp_getattro, (void *)PyObject_GenericGetAttr },
+    { Py_tp_doc, (void *)SwigPyPacked_doc },
+    { 0, NULL }
+  };
+  PyType_Spec spec = {
+    SWIG_RUNTIME_MODULE ".SwigPyPacked",
+    sizeof(SwigPyPacked),
+    0,
+    Py_TPFLAGS_DEFAULT,
+    slots
+  };
+  PyObject *pytype = PyType_FromSpec(&spec);
+  PyObject *runtime_data_module = SWIG_runtime_data_module();
+  if (pytype && PyModule_AddObject(runtime_data_module, "SwigPyPacked", pytype) == 0)
+    SWIG_Py_INCREF(pytype);
+  return (PyTypeObject *)pytype;
+#endif
 }
 
 SWIGRUNTIME PyObject *
 SwigPyPacked_New(void *ptr, size_t size, swig_type_info *ty)
 {
-  SwigPyPacked *sobj = PyObject_NEW(SwigPyPacked, SwigPyPacked_type());
+  SwigPyPacked *sobj = PyObject_New(SwigPyPacked, SwigPyPacked_Type());
   if (sobj) {
     void *pack = malloc(size);
     if (pack) {
@@ -2246,7 +2642,7 @@ SwigPyPacked_New(void *ptr, size_t size, swig_type_info *ty)
       sobj->ty   = ty;
       sobj->size = size;
     } else {
-      PyObject_DEL((PyObject *) sobj);
+      PyObject_Free((PyObject *)sobj);
       sobj = 0;
     }
   }
@@ -2283,7 +2679,7 @@ SWIG_This(void)
 /* #define SWIG_PYTHON_SLOW_GETSET_THIS */
 
 /* TODO: I don't know how to implement the fast getset in Python 3 right now */
-#if PY_VERSION_HEX>=0x03000000
+#if PY_VERSION_HEX >= 0x03000000
 #define SWIG_PYTHON_SLOW_GETSET_THIS 
 #endif
 
@@ -2299,7 +2695,12 @@ SWIG_Python_GetSwigThis(PyObject *pyobj)
   (void)obj;
 # ifdef PyWeakref_CheckProxy
   if (PyWeakref_CheckProxy(pyobj)) {
+#if PY_VERSION_HEX >= 0x030d0000
+    PyWeakref_GetRef(pyobj, &pyobj);
+    Py_DECREF(pyobj);
+#else
     pyobj = PyWeakref_GET_OBJECT(pyobj);
+#endif
     if (pyobj && SwigPyObject_Check(pyobj))
       return (SwigPyObject*) pyobj;
   }
@@ -2326,7 +2727,7 @@ SWIG_Python_GetSwigThis(PyObject *pyobj)
 #endif
       obj = PyObject_GetAttr(pyobj,SWIG_This());
       if (obj) {
-	Py_DECREF(obj);
+	SWIG_Py_DECREF(obj);
       } else {
 	if (PyErr_Occurred()) PyErr_Clear();
 	return 0;
@@ -2336,7 +2737,7 @@ SWIG_Python_GetSwigThis(PyObject *pyobj)
 #else
   obj = PyObject_GetAttr(pyobj,SWIG_This());
   if (obj) {
-    Py_DECREF(obj);
+    SWIG_Py_DECREF(obj);
   } else {
     if (PyErr_Occurred()) PyErr_Clear();
     return 0;
@@ -2462,7 +2863,7 @@ SWIG_Python_ConvertPtrAndOwn(PyObject *obj, void **ptr, swig_type_info *ty, int 
                 }
               }
             }
-            Py_DECREF(impconv);
+            SWIG_Py_DECREF(impconv);
           }
         }
       }
@@ -2489,10 +2890,20 @@ SWIG_Python_ConvertFunctionPtr(PyObject *obj, void **ptr, swig_type_info *ty) {
     swig_cast_info *tc;
 
     /* here we get the method pointer for callbacks */
+#ifndef Py_LIMITED_API
     const char *doc = (((PyCFunctionObject *)obj) -> m_ml -> ml_doc);
+#else
+    PyObject* pystr_doc = PyObject_GetAttrString(obj, "__doc__");
+    PyObject *bytes = NULL;
+    const char *doc = pystr_doc ? SWIG_PyUnicode_AsUTF8AndSize(pystr_doc, NULL, &bytes) : 0;
+#endif
     const char *desc = doc ? strstr(doc, "swig_ptr: ") : 0;
     if (desc)
       desc = ty ? SWIG_UnpackVoidPtr(desc + 10, &vptr, ty->name) : 0;
+#ifdef Py_LIMITED_API
+    SWIG_Py_XDECREF(bytes);
+    SWIG_Py_XDECREF(pystr_doc);
+#endif
     if (!desc)
       return SWIG_ERROR;
     tc = SWIG_TypeCheck(desc,ty);
@@ -2551,13 +2962,13 @@ SWIG_Python_NewShadowInstance(SwigPyClientData *data, PyObject *swig_this)
         if (dict) {
           PyDict_SetItem(dict, SWIG_This(), swig_this);
         } else{
-          Py_DECREF(inst);
+          SWIG_Py_DECREF(inst);
           inst = 0;
         }
       }
 #else
       if (PyObject_SetAttr(inst, SWIG_This(), swig_this) == -1) {
-        Py_DECREF(inst);
+        SWIG_Py_DECREF(inst);
         inst = 0;
       }
 #endif
@@ -2568,25 +2979,30 @@ SWIG_Python_NewShadowInstance(SwigPyClientData *data, PyObject *swig_this)
     if (empty_args) {
       PyObject *empty_kwargs = PyDict_New();
       if (empty_kwargs) {
-        inst = ((PyTypeObject *)data->newargs)->tp_new((PyTypeObject *)data->newargs, empty_args, empty_kwargs);
-        Py_DECREF(empty_kwargs);
+#ifndef Py_LIMITED_API
+        newfunc newfn = ((PyTypeObject *)data->newargs)->tp_new;
+#else
+        newfunc newfn = (newfunc)PyType_GetSlot((PyTypeObject *)data->newargs, Py_tp_new);
+#endif
+        inst = newfn((PyTypeObject *)data->newargs, empty_args, empty_kwargs);
+        SWIG_Py_DECREF(empty_kwargs);
         if (inst) {
           if (PyObject_SetAttr(inst, SWIG_This(), swig_this) == -1) {
-            Py_DECREF(inst);
+            SWIG_Py_DECREF(inst);
             inst = 0;
           } else {
             PyType_Modified(Py_TYPE(inst));
           }
         }
       }
-      Py_DECREF(empty_args);
+      SWIG_Py_DECREF(empty_args);
     }
 #else
     PyObject *dict = PyDict_New();
     if (dict) {
       PyDict_SetItem(dict, SWIG_This(), swig_this);
       inst = PyInstance_NewRaw(data->newargs, dict);
-      Py_DECREF(dict);
+      SWIG_Py_DECREF(dict);
     }
 #endif
   }
@@ -2623,7 +3039,7 @@ SWIG_Python_InitShadowInstance(PyObject *args) {
   } else {
     SwigPyObject *sthis = SWIG_Python_GetSwigThis(obj[0]);
     if (sthis) {
-      Py_DECREF(SwigPyObject_append((PyObject*) sthis, obj[1]));
+      SWIG_Py_DECREF(SwigPyObject_append((PyObject*) sthis, obj[1]));
     } else {
       if (SWIG_Python_SetSwigThis(obj[0], obj[1]) != 0)
         return NULL;
@@ -2650,22 +3066,23 @@ SWIG_Python_NewPointerObj(PyObject *self, void *ptr, swig_type_info *type, int f
     if (flags & SWIG_BUILTIN_TP_INIT) {
       newobj = (SwigPyObject*) self;
       if (newobj->ptr) {
-        PyObject *next_self = clientdata->pytype->tp_alloc(clientdata->pytype, 0);
+#ifndef Py_LIMITED_API
+        allocfunc alloc = clientdata->pytype->tp_alloc;
+#else
+        allocfunc alloc = (allocfunc)PyType_GetSlot(clientdata->pytype, Py_tp_alloc);
+#endif
+        PyObject *next_self = alloc(clientdata->pytype, 0);
         while (newobj->next)
 	  newobj = (SwigPyObject *) newobj->next;
         newobj->next = next_self;
         newobj = (SwigPyObject *)next_self;
-#ifdef SWIGPYTHON_BUILTIN
-        newobj->dict = 0;
-#endif
+        newobj->swigdict = 0;
       }
     } else {
       newobj = PyObject_New(SwigPyObject, clientdata->pytype);
-#ifdef SWIGPYTHON_BUILTIN
       if (newobj) {
-        newobj->dict = 0;
+        newobj->swigdict = 0;
       }
-#endif
     }
     if (newobj) {
       newobj->ptr = ptr;
@@ -2682,7 +3099,7 @@ SWIG_Python_NewPointerObj(PyObject *self, void *ptr, swig_type_info *type, int f
   robj = SwigPyObject_New(ptr, type, own);
   if (robj && clientdata && !(flags & SWIG_POINTER_NOSHADOW)) {
     PyObject *inst = SWIG_Python_NewShadowInstance(clientdata, robj);
-    Py_DECREF(robj);
+    SWIG_Py_DECREF(robj);
     robj = inst;
   }
   return robj;
@@ -2733,7 +3150,13 @@ SWIG_Python_GetModule(void *SWIGUNUSEDPARM(clientdata)) {
 }
 
 
-static int interpreter_counter = 0; // how many (sub-)interpreters are using swig_module's types
+#if defined(SWIG_REFCNT_DEBUG)
+#define SWIG_PYOBJ_REFCNT(OBJ) fprintf(stdout, "" #OBJ " count %ld\n", (OBJ ? Py_REFCNT(OBJ) : 0))
+#else
+#define SWIG_PYOBJ_REFCNT(OBJ)
+#endif
+
+static int interpreter_counter = 0; /* how many (sub-)interpreters are using swig_module's types */
 
 SWIGRUNTIME void
 SWIG_Python_DestroyModule(PyObject *obj)
@@ -2741,9 +3164,9 @@ SWIG_Python_DestroyModule(PyObject *obj)
   swig_module_info *swig_module = (swig_module_info *) PyCapsule_GetPointer(obj, SWIGPY_CAPSULE_NAME);
   swig_type_info **types = swig_module->types;
   size_t i;
-  if (--interpreter_counter != 0) // another sub-interpreter may still be using the swig_module's types
+  if (--interpreter_counter != 0) /* another sub-interpreter may still be using the swig_module's types */
     return;
-  for (i =0; i < swig_module->size; ++i) {
+  for (i = 0; i < swig_module->size; ++i) {
     swig_type_info *ty = types[i];
     if (ty->owndata) {
       SwigPyClientData *data = (SwigPyClientData *) ty->clientdata;
@@ -2751,34 +3174,39 @@ SWIG_Python_DestroyModule(PyObject *obj)
       if (data) SwigPyClientData_Del(data);
     }
   }
-  Py_DECREF(SWIG_This());
+  SWIG_PYOBJ_REFCNT(Swig_This_global);
+  SWIG_Py_XDECREF(Swig_This_global);
   Swig_This_global = NULL;
-  Py_DECREF(SWIG_globals());
+
+  SWIG_PYOBJ_REFCNT(Swig_Globals_global);
+  SWIG_Py_XDECREF(Swig_Globals_global);
   Swig_Globals_global = NULL;
-  Py_DECREF(SWIG_Python_TypeCache());
+
+  SWIG_PYOBJ_REFCNT(Swig_TypeCache_global);
+  SWIG_Py_XDECREF(Swig_TypeCache_global);
   Swig_TypeCache_global = NULL;
+
+  SWIG_PYOBJ_REFCNT(Swig_Capsule_global);
   Swig_Capsule_global = NULL;
+
+  SWIG_PYOBJ_REFCNT(Swig_runtime_data_module_global);
+  SWIG_Py_XDECREF(Swig_runtime_data_module_global);
+  Swig_runtime_data_module_global = NULL;
 }
 
 SWIGRUNTIME void
 SWIG_Python_SetModule(swig_module_info *swig_module) {
-#if PY_VERSION_HEX >= 0x03000000
- /* Add a dummy module object into sys.modules */
-  PyObject *module = PyImport_AddModule("swig_runtime_data" SWIG_RUNTIME_VERSION);
-#else
-  static PyMethodDef swig_empty_runtime_method_table[] = { {NULL, NULL, 0, NULL} }; /* Sentinel */
-  PyObject *module = Py_InitModule("swig_runtime_data" SWIG_RUNTIME_VERSION, swig_empty_runtime_method_table);
-#endif
+  PyObject *runtime_data_module = SWIG_runtime_data_module();
   PyObject *pointer = PyCapsule_New((void *) swig_module, SWIGPY_CAPSULE_NAME, SWIG_Python_DestroyModule);
-  if (pointer && module) {
-    if (PyModule_AddObject(module, SWIGPY_CAPSULE_ATTR_NAME, pointer) == 0) {
+  if (pointer && runtime_data_module) {
+    if (PyModule_AddObject(runtime_data_module, SWIGPY_CAPSULE_ATTR_NAME, pointer) == 0) {
       ++interpreter_counter;
       Swig_Capsule_global = pointer;
     } else {
-      Py_DECREF(pointer);
+      SWIG_Py_DECREF(pointer);
     }
   } else {
-    Py_XDECREF(pointer);
+    SWIG_Py_XDECREF(pointer);
   }
 }
 
@@ -2798,11 +3226,11 @@ SWIG_Python_TypeQuery(const char *type)
       obj = PyCapsule_New((void*) descriptor, NULL, NULL);
       if (obj) {
         PyDict_SetItem(cache, key, obj);
-        Py_DECREF(obj);
+        SWIG_Py_DECREF(obj);
       }
     }
   }
-  Py_DECREF(key);
+  SWIG_Py_DECREF(key);
   return descriptor;
 }
 
@@ -2823,16 +3251,18 @@ SWIG_Python_AddErrMesg(const char* mesg, int infront)
     PyErr_Fetch(&type, &value, &traceback);
     if (value) {
       PyObject *old_str = PyObject_Str(value);
-      const char *tmp = SWIG_Python_str_AsChar(old_str);
+      PyObject *bytes = NULL;
+      const char *tmp = SWIG_PyUnicode_AsUTF8AndSize(old_str, NULL, &bytes);
       const char *errmesg = tmp ? tmp : "Invalid error message";
-      Py_XINCREF(type);
+      SWIG_Py_XINCREF(type);
       PyErr_Clear();
       if (infront) {
 	PyErr_Format(type, "%s %s", mesg, errmesg);
       } else {
 	PyErr_Format(type, "%s %s", errmesg, mesg);
       }
-      Py_DECREF(old_str);
+      SWIG_Py_XDECREF(bytes);
+      SWIG_Py_DECREF(old_str);
     }
     return 1;
   } else {
@@ -2861,43 +3291,6 @@ SwigPyObject_GetDesc(PyObject *self)
   return ty ? ty->str : "";
 }
 
-SWIGRUNTIME void
-SWIG_Python_TypeError(const char *type, PyObject *obj)
-{
-  if (type) {
-#if defined(SWIG_COBJECT_TYPES)
-    if (obj && SwigPyObject_Check(obj)) {
-      const char *otype = (const char *) SwigPyObject_GetDesc(obj);
-      if (otype) {
-	PyErr_Format(PyExc_TypeError, "a '%s' is expected, 'SwigPyObject(%s)' is received",
-		     type, otype);
-	return;
-      }
-    } else 
-#endif      
-    {
-      const char *otype = (obj ? obj->ob_type->tp_name : 0); 
-      if (otype) {
-	PyObject *str = PyObject_Str(obj);
-	const char *cstr = str ? SWIG_Python_str_AsChar(str) : 0;
-	if (cstr) {
-	  PyErr_Format(PyExc_TypeError, "a '%s' is expected, '%s(%s)' is received",
-		       type, otype, cstr);
-	} else {
-	  PyErr_Format(PyExc_TypeError, "a '%s' is expected, '%s' is received",
-		       type, otype);
-	}
-	Py_XDECREF(str);
-	return;
-      }
-    }   
-    PyErr_Format(PyExc_TypeError, "a '%s' is expected", type);
-  } else {
-    PyErr_Format(PyExc_TypeError, "unexpected type is received");
-  }
-}
-
-
 /* Convert a pointer value, signal an exception on a type mismatch */
 SWIGRUNTIME void *
 SWIG_Python_MustGetPtr(PyObject *obj, swig_type_info *ty, int SWIGUNUSEDPARM(argnum), int flags) {
@@ -2911,7 +3304,7 @@ SWIG_Python_MustGetPtr(PyObject *obj, swig_type_info *ty, int SWIGUNUSEDPARM(arg
 #ifdef SWIGPYTHON_BUILTIN
 SWIGRUNTIME int
 SWIG_Python_NonDynamicSetAttr(PyObject *obj, PyObject *name, PyObject *value) {
-  PyTypeObject *tp = obj->ob_type;
+  PyTypeObject *tp = Py_TYPE(obj);
   PyObject *descr;
   PyObject *encoded_name;
   descrsetfunc f;
@@ -2927,10 +3320,16 @@ SWIG_Python_NonDynamicSetAttr(PyObject *obj, PyObject *name, PyObject *value) {
   if (!PyString_Check(name))
 # endif
   {
-    PyErr_Format(PyExc_TypeError, "attribute name must be string, not '%.200s'", name->ob_type->tp_name);
+#if PY_VERSION_HEX >= 0x03000000
+    PyObject *tpname = SWIG_PyType_GetFullyQualifiedName(Py_TYPE(name));
+    PyErr_Format(PyExc_TypeError, "attribute name must be string, not '%S'", tpname);
+    SWIG_Py_DECREF(tpname);
+#else
+    PyErr_Format(PyExc_TypeError, "attribute name must be string, not '%s'", Py_TYPE(name)->tp_name);
+#endif
     return -1;
   } else {
-    Py_INCREF(name);
+    SWIG_Py_INCREF(name);
   }
 
   if (!tp->tp_dict) {
@@ -2941,24 +3340,32 @@ SWIG_Python_NonDynamicSetAttr(PyObject *obj, PyObject *name, PyObject *value) {
   descr = _PyType_Lookup(tp, name);
   f = NULL;
   if (descr != NULL)
-    f = descr->ob_type->tp_descr_set;
+    f = Py_TYPE(descr)->tp_descr_set;
   if (!f) {
     if (PyString_Check(name)) {
       encoded_name = name;
-      Py_INCREF(name);
+      SWIG_Py_INCREF(name);
     } else {
       encoded_name = PyUnicode_AsUTF8String(name);
       if (!encoded_name)
         goto done;
     }
-    PyErr_Format(PyExc_AttributeError, "'%.100s' object has no attribute '%.200s'", tp->tp_name, PyString_AsString(encoded_name));
-    Py_DECREF(encoded_name);
+#if PY_VERSION_HEX >= 0x03000000
+    {
+      PyObject *tpname = SWIG_PyType_GetFullyQualifiedName(tp);
+      PyErr_Format(PyExc_AttributeError, "'%S' object has no attribute '%s'", tpname, PyString_AsString(encoded_name));
+      SWIG_Py_DECREF(tpname);
+    }
+#else
+    PyErr_Format(PyExc_AttributeError, "'%s' object has no attribute '%s'", tp->tp_name, PyString_AsString(encoded_name));
+#endif
+    SWIG_Py_DECREF(encoded_name);
   } else {
     res = f(descr, obj, value);
   }
   
   done:
-  Py_DECREF(name);
+  SWIG_Py_DECREF(name);
   return res;
 }
 #endif
@@ -3018,7 +3425,6 @@ static swig_module_info swig_module = {swig_types, 15, 0, 0, 0, 0};
 #  define SWIG_init    init_sentencepiece
 
 #endif
-#define SWIG_name    "_sentencepiece"
 
 #ifdef __cplusplus
 #include <utility>
@@ -3088,7 +3494,7 @@ namespace swig {
     SwigPtr_PyObject(const SwigPtr_PyObject& item) : _obj(item._obj)
     {
       SWIG_PYTHON_THREAD_BEGIN_BLOCK;
-      Py_XINCREF(_obj);      
+      SWIG_Py_XINCREF(_obj);
       SWIG_PYTHON_THREAD_END_BLOCK;
     }
     
@@ -3096,7 +3502,7 @@ namespace swig {
     {
       if (initial_ref) {
         SWIG_PYTHON_THREAD_BEGIN_BLOCK;
-        Py_XINCREF(_obj);
+        SWIG_Py_XINCREF(_obj);
         SWIG_PYTHON_THREAD_END_BLOCK;
       }
     }
@@ -3104,8 +3510,8 @@ namespace swig {
     SwigPtr_PyObject & operator=(const SwigPtr_PyObject& item) 
     {
       SWIG_PYTHON_THREAD_BEGIN_BLOCK;
-      Py_XINCREF(item._obj);
-      Py_XDECREF(_obj);
+      SWIG_Py_XINCREF(item._obj);
+      SWIG_Py_XDECREF(_obj);
       _obj = item._obj;
       SWIG_PYTHON_THREAD_END_BLOCK;
       return *this;      
@@ -3114,7 +3520,7 @@ namespace swig {
     ~SwigPtr_PyObject() 
     {
       SWIG_PYTHON_THREAD_BEGIN_BLOCK;
-      Py_XDECREF(_obj);
+      SWIG_Py_XDECREF(_obj);
       SWIG_PYTHON_THREAD_END_BLOCK;
     }
     
@@ -3137,7 +3543,7 @@ namespace swig {
     
     SwigVar_PyObject & operator = (PyObject* obj)
     {
-      Py_XDECREF(_obj);
+      SWIG_Py_XDECREF(_obj);
       _obj = obj;
       return *this;      
     }
@@ -3543,6 +3949,9 @@ SWIG_AsVal_double (PyObject *obj, double *val)
 }
 
 
+#include <errno.h>
+
+
 #include <float.h>
 
 
@@ -3553,9 +3962,11 @@ SWIGINTERNINLINE int
 SWIG_CanCastAsInteger(double *d, double min, double max) {
   double x = *d;
   if ((min <= x && x <= max)) {
-   double fx = floor(x);
-   double cx = ceil(x);
-   double rd =  ((x - fx) < 0.5) ? fx : cx; /* simple rint */
+   double fx, cx, rd;
+   errno = 0;
+   fx = floor(x);
+   cx = ceil(x);
+   rd =  ((x - fx) < 0.5) ? fx : cx; /* simple rint */
    if ((errno == EDOM) || (errno == ERANGE)) {
      errno = 0;
    } else {
@@ -3611,7 +4022,12 @@ SWIG_AsVal_long (PyObject *obj, long* val)
     if (!dispatch) {
       double d;
       int res = SWIG_AddCast(SWIG_AsVal_double (obj,&d));
-      if (SWIG_IsOK(res) && SWIG_CanCastAsInteger(&d, LONG_MIN, LONG_MAX)) {
+      // Largest double not larger than LONG_MAX (not portably calculated easily)
+      // Note that double(LONG_MAX) is stored in a double rounded up by one (for 64-bit long)
+      // 0x7ffffffffffffc00LL == (int64_t)std::nextafter(double(__uint128_t(LONG_MAX)+1), double(0))
+      const double long_max = sizeof(long) == 8 ? 0x7ffffffffffffc00LL : LONG_MAX;
+      // No equivalent needed for 64-bit double(LONG_MIN) is exactly LONG_MIN
+      if (SWIG_IsOK(res) && SWIG_CanCastAsInteger(&d, LONG_MIN, long_max)) {
 	if (val) *val = (long)(d);
 	return res;
       }
@@ -3654,20 +4070,20 @@ SWIGINTERN sentencepiece::util::bytes const &sentencepiece_ImmutableSentencePiec
 SWIGINTERN swig_type_info*
 SWIG_pchar_descriptor(void)
 {
-  static int init = 0;
   static swig_type_info* info = 0;
-  if (!init) {
+  if (!info) {
     info = SWIG_TypeQuery("_p_char");
-    init = 1;
   }
   return info;
 }
 
 
+/* Return string from Python obj. NOTE: obj must remain in scope in order
+   to use the returned cptr (but only when alloc is set to SWIG_OLDOBJ) */
 SWIGINTERN int
-SWIG_AsCharPtrAndSize(PyObject *obj, char** cptr, size_t* psize, int *alloc)
+SWIG_AsCharPtrAndSize(PyObject *obj, char **cptr, size_t *psize, int *alloc)
 {
-#if PY_VERSION_HEX>=0x03000000
+#if PY_VERSION_HEX >= 0x03000000
 #if defined(SWIG_PYTHON_STRICT_BYTE_CHAR)
   if (PyBytes_Check(obj))
 #else
@@ -3678,62 +4094,38 @@ SWIG_AsCharPtrAndSize(PyObject *obj, char** cptr, size_t* psize, int *alloc)
 #endif
   {
     char *cstr; Py_ssize_t len;
+    PyObject *bytes = NULL;
     int ret = SWIG_OK;
-#if PY_VERSION_HEX>=0x03000000
-#if !defined(SWIG_PYTHON_STRICT_BYTE_CHAR)
-    if (!alloc && cptr) {
-        /* We can't allow converting without allocation, since the internal
-           representation of string in Python 3 is UCS-2/UCS-4 but we require
-           a UTF-8 representation.
-           TODO(bhy) More detailed explanation */
-        return SWIG_RuntimeError;
-    }
-    obj = PyUnicode_AsUTF8String(obj);
-    if (!obj)
-      return SWIG_TypeError;
     if (alloc)
-      *alloc = SWIG_NEWOBJ;
-#endif
+      *alloc = SWIG_OLDOBJ;
+#if PY_VERSION_HEX >= 0x03000000 && defined(SWIG_PYTHON_STRICT_BYTE_CHAR)
     if (PyBytes_AsStringAndSize(obj, &cstr, &len) == -1)
       return SWIG_TypeError;
 #else
-    if (PyString_AsStringAndSize(obj, &cstr, &len) == -1)
+    cstr = (char *)SWIG_PyUnicode_AsUTF8AndSize(obj, &len, &bytes);
+    if (!cstr)
       return SWIG_TypeError;
-#endif
-    if (cptr) {
+    /* The returned string is only duplicated if the char * returned is not owned and memory managed by obj */
+    if (bytes && cptr) {
       if (alloc) {
-	if (*alloc == SWIG_NEWOBJ) {
-	  *cptr = reinterpret_cast< char* >(memcpy(new char[len + 1], cstr, sizeof(char)*(len + 1)));
-	  *alloc = SWIG_NEWOBJ;
-	} else {
-	  *cptr = cstr;
-	  *alloc = SWIG_OLDOBJ;
-	}
+        cstr = reinterpret_cast< char* >(memcpy(new char[len + 1], cstr, sizeof(char)*(len + 1)));
+        *alloc = SWIG_NEWOBJ;
       } else {
-#if PY_VERSION_HEX>=0x03000000
-#if defined(SWIG_PYTHON_STRICT_BYTE_CHAR)
-	*cptr = PyBytes_AsString(obj);
-#else
-	assert(0); /* Should never reach here with Unicode strings in Python 3 */
-#endif
-#else
-	*cptr = SWIG_Python_str_AsChar(obj);
-        if (!*cptr)
-          ret = SWIG_TypeError;
-#endif
+        /* alloc must be set in order to clean up allocated memory */
+        return SWIG_RuntimeError;
       }
     }
-    if (psize) *psize = len + 1;
-#if PY_VERSION_HEX>=0x03000000 && !defined(SWIG_PYTHON_STRICT_BYTE_CHAR)
-    Py_XDECREF(obj);
 #endif
+    if (cptr) *cptr = cstr;
+    if (psize) *psize = len + 1;
+    SWIG_Py_XDECREF(bytes);
     return ret;
   } else {
 #if defined(SWIG_PYTHON_2_UNICODE)
 #if defined(SWIG_PYTHON_STRICT_BYTE_CHAR)
 #error "Cannot use both SWIG_PYTHON_2_UNICODE and SWIG_PYTHON_STRICT_BYTE_CHAR at once"
 #endif
-#if PY_VERSION_HEX<0x03000000
+#if PY_VERSION_HEX < 0x03000000
     if (PyUnicode_Check(obj)) {
       char *cstr; Py_ssize_t len;
       if (!alloc && cptr) {
@@ -3749,10 +4141,10 @@ SWIG_AsCharPtrAndSize(PyObject *obj, char** cptr, size_t* psize, int *alloc)
         }
         if (psize) *psize = len + 1;
 
-        Py_XDECREF(obj);
+        SWIG_Py_XDECREF(obj);
         return SWIG_OK;
       } else {
-        Py_XDECREF(obj);
+        SWIG_Py_XDECREF(obj);
       }
     }
 #endif
@@ -4097,7 +4489,11 @@ SWIG_AsVal_unsigned_SS_long (PyObject *obj, unsigned long *val)
     if (!dispatch) {
       double d;
       int res = SWIG_AddCast(SWIG_AsVal_double (obj,&d));
-      if (SWIG_IsOK(res) && SWIG_CanCastAsInteger(&d, 0, ULONG_MAX)) {
+      // Largest double not larger than ULONG_MAX (not portably calculated easily)
+      // Note that double(ULONG_MAX) is stored in a double rounded up by one (for 64-bit unsigned long)
+      // 0xfffffffffffff800ULL == (uint64_t)std::nextafter(double(__uint128_t(ULONG_MAX)+1), double(0))
+      const double ulong_max = sizeof(unsigned long) == 8 ? 0xfffffffffffff800ULL : ULONG_MAX;
+      if (SWIG_IsOK(res) && SWIG_CanCastAsInteger(&d, 0, ulong_max)) {
 	if (val) *val = (unsigned long)(d);
 	return res;
       }
@@ -4178,6 +4574,7 @@ SWIGINTERN PyObject *_wrap_new_ImmutableSentencePieceText_ImmutableSentencePiece
   PyObject *resultobj = 0;
   sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *result = 0 ;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "new_ImmutableSentencePieceText_ImmutableSentencePiece", 0, 0, 0)) SWIG_fail;
   {
     try {
@@ -4197,11 +4594,12 @@ fail:
 
 SWIGINTERN PyObject *_wrap_delete_ImmutableSentencePieceText_ImmutableSentencePiece(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = (sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *) 0 ;
+  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, SWIG_POINTER_DISOWN |  0 );
@@ -4227,12 +4625,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText_ImmutableSentencePiece__piece(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = (sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *) 0 ;
+  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   std::string *result = 0 ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, 0 |  0 );
@@ -4261,12 +4660,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText_ImmutableSentencePiece__surface(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = (sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *) 0 ;
+  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   std::string *result = 0 ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, 0 |  0 );
@@ -4295,12 +4695,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText_ImmutableSentencePiece__id(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = (sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *) 0 ;
+  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   uint32_t result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, 0 |  0 );
@@ -4326,12 +4727,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText_ImmutableSentencePiece__begin(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = (sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *) 0 ;
+  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   uint32_t result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, 0 |  0 );
@@ -4357,12 +4759,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText_ImmutableSentencePiece__end(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = (sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *) 0 ;
+  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   uint32_t result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, 0 |  0 );
@@ -4388,12 +4791,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText_ImmutableSentencePiece__surface_as_bytes(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = (sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *) 0 ;
+  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   sentencepiece::util::bytes *result = 0 ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, 0 |  0 );
@@ -4421,12 +4825,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText_ImmutableSentencePiece__piece_as_bytes(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = (sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *) 0 ;
+  sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   sentencepiece::util::bytes *result = 0 ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, 0 |  0 );
@@ -4453,7 +4858,7 @@ fail:
 
 
 SWIGINTERN PyObject *ImmutableSentencePieceText_ImmutableSentencePiece_swigregister(PyObject *SWIGUNUSEDPARM(self), PyObject *args) {
-  PyObject *obj;
+  PyObject *obj = NULL;
   if (!SWIG_Python_UnpackTuple(args, "swigregister", 1, 1, &obj)) return NULL;
   SWIG_TypeNewClientData(SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText_ImmutableSentencePiece, SWIG_NewClientData(obj));
   return SWIG_Py_Void();
@@ -4467,6 +4872,7 @@ SWIGINTERN PyObject *_wrap_new_ImmutableSentencePieceText(PyObject *self, PyObje
   PyObject *resultobj = 0;
   sentencepiece::ImmutableSentencePieceText *result = 0 ;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "new_ImmutableSentencePieceText", 0, 0, 0)) SWIG_fail;
   {
     try {
@@ -4486,11 +4892,12 @@ fail:
 
 SWIGINTERN PyObject *_wrap_delete_ImmutableSentencePieceText(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText *arg1 = (sentencepiece::ImmutableSentencePieceText *) 0 ;
+  sentencepiece::ImmutableSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText, SWIG_POINTER_DISOWN |  0 );
@@ -4516,12 +4923,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText__pieces_size(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText *arg1 = (sentencepiece::ImmutableSentencePieceText *) 0 ;
+  sentencepiece::ImmutableSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   size_t result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText, 0 |  0 );
@@ -4547,7 +4955,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText__pieces(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText *arg1 = (sentencepiece::ImmutableSentencePieceText *) 0 ;
+  sentencepiece::ImmutableSentencePieceText *arg1 = 0 ;
   int arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
@@ -4556,6 +4964,7 @@ SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText__pieces(PyObject *self, Py
   PyObject *swig_obj[2] ;
   sentencepiece::ImmutableSentencePieceText_ImmutableSentencePiece result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "ImmutableSentencePieceText__pieces", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -4585,12 +4994,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText__text(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText *arg1 = (sentencepiece::ImmutableSentencePieceText *) 0 ;
+  sentencepiece::ImmutableSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   std::string *result = 0 ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText, 0 |  0 );
@@ -4619,12 +5029,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText__score(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText *arg1 = (sentencepiece::ImmutableSentencePieceText *) 0 ;
+  sentencepiece::ImmutableSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   float result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText, 0 |  0 );
@@ -4650,12 +5061,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText_SerializeAsString(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText *arg1 = (sentencepiece::ImmutableSentencePieceText *) 0 ;
+  sentencepiece::ImmutableSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText, 0 |  0 );
@@ -4683,12 +5095,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableSentencePieceText__text_as_bytes(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableSentencePieceText *arg1 = (sentencepiece::ImmutableSentencePieceText *) 0 ;
+  sentencepiece::ImmutableSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   sentencepiece::util::bytes *result = 0 ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText, 0 |  0 );
@@ -4715,7 +5128,7 @@ fail:
 
 
 SWIGINTERN PyObject *ImmutableSentencePieceText_swigregister(PyObject *SWIGUNUSEDPARM(self), PyObject *args) {
-  PyObject *obj;
+  PyObject *obj = NULL;
   if (!SWIG_Python_UnpackTuple(args, "swigregister", 1, 1, &obj)) return NULL;
   SWIG_TypeNewClientData(SWIGTYPE_p_sentencepiece__ImmutableSentencePieceText, SWIG_NewClientData(obj));
   return SWIG_Py_Void();
@@ -4729,6 +5142,7 @@ SWIGINTERN PyObject *_wrap_new_ImmutableNBestSentencePieceText(PyObject *self, P
   PyObject *resultobj = 0;
   sentencepiece::ImmutableNBestSentencePieceText *result = 0 ;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "new_ImmutableNBestSentencePieceText", 0, 0, 0)) SWIG_fail;
   {
     try {
@@ -4748,11 +5162,12 @@ fail:
 
 SWIGINTERN PyObject *_wrap_delete_ImmutableNBestSentencePieceText(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableNBestSentencePieceText *arg1 = (sentencepiece::ImmutableNBestSentencePieceText *) 0 ;
+  sentencepiece::ImmutableNBestSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableNBestSentencePieceText, SWIG_POINTER_DISOWN |  0 );
@@ -4778,12 +5193,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableNBestSentencePieceText__nbests_size(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableNBestSentencePieceText *arg1 = (sentencepiece::ImmutableNBestSentencePieceText *) 0 ;
+  sentencepiece::ImmutableNBestSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   size_t result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableNBestSentencePieceText, 0 |  0 );
@@ -4809,7 +5225,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableNBestSentencePieceText__nbests(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableNBestSentencePieceText *arg1 = (sentencepiece::ImmutableNBestSentencePieceText *) 0 ;
+  sentencepiece::ImmutableNBestSentencePieceText *arg1 = 0 ;
   int arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
@@ -4818,6 +5234,7 @@ SWIGINTERN PyObject *_wrap_ImmutableNBestSentencePieceText__nbests(PyObject *sel
   PyObject *swig_obj[2] ;
   sentencepiece::ImmutableSentencePieceText result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "ImmutableNBestSentencePieceText__nbests", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableNBestSentencePieceText, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -4847,12 +5264,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_ImmutableNBestSentencePieceText_SerializeAsString(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::ImmutableNBestSentencePieceText *arg1 = (sentencepiece::ImmutableNBestSentencePieceText *) 0 ;
+  sentencepiece::ImmutableNBestSentencePieceText *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__ImmutableNBestSentencePieceText, 0 |  0 );
@@ -4879,7 +5297,7 @@ fail:
 
 
 SWIGINTERN PyObject *ImmutableNBestSentencePieceText_swigregister(PyObject *SWIGUNUSEDPARM(self), PyObject *args) {
-  PyObject *obj;
+  PyObject *obj = NULL;
   if (!SWIG_Python_UnpackTuple(args, "swigregister", 1, 1, &obj)) return NULL;
   SWIG_TypeNewClientData(SWIGTYPE_p_sentencepiece__ImmutableNBestSentencePieceText, SWIG_NewClientData(obj));
   return SWIG_Py_Void();
@@ -4893,6 +5311,7 @@ SWIGINTERN PyObject *_wrap_new_SentencePieceProcessor(PyObject *self, PyObject *
   PyObject *resultobj = 0;
   sentencepiece::SentencePieceProcessor *result = 0 ;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "new_SentencePieceProcessor", 0, 0, 0)) SWIG_fail;
   {
     try {
@@ -4912,11 +5331,12 @@ fail:
 
 SWIGINTERN PyObject *_wrap_delete_SentencePieceProcessor(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, SWIG_POINTER_DISOWN |  0 );
@@ -4942,13 +5362,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_LoadFromSerializedProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_LoadFromSerializedProto", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -4966,7 +5387,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_LoadFromSerializedProto(PyObje
   }
   {
     try {
-      result = (arg1)->LoadFromSerializedProto(arg2);
+      result = (arg1)->LoadFromSerializedProto(SWIG_STD_MOVE(arg2));
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -4987,13 +5408,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_SetEncodeExtraOptions(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_SetEncodeExtraOptions", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5011,7 +5433,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_SetEncodeExtraOptions(PyObject
   }
   {
     try {
-      result = (arg1)->SetEncodeExtraOptions(arg2);
+      result = (arg1)->SetEncodeExtraOptions(SWIG_STD_MOVE(arg2));
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -5032,13 +5454,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_SetDecodeExtraOptions(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_SetDecodeExtraOptions", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5056,7 +5479,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_SetDecodeExtraOptions(PyObject
   }
   {
     try {
-      result = (arg1)->SetDecodeExtraOptions(arg2);
+      result = (arg1)->SetDecodeExtraOptions(SWIG_STD_MOVE(arg2));
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -5077,13 +5500,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_SetVocabulary(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_SetVocabulary", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5140,12 +5564,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_ResetVocabulary(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
@@ -5176,7 +5601,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_LoadVocabulary(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   void *argp1 = 0 ;
@@ -5186,6 +5611,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_LoadVocabulary(PyObject *self,
   PyObject *swig_obj[3] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_LoadVocabulary", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5208,7 +5634,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_LoadVocabulary(PyObject *self,
   arg3 = static_cast< int >(val3);
   {
     try {
-      result = (arg1)->LoadVocabulary(arg2,arg3);
+      result = (arg1)->LoadVocabulary(SWIG_STD_MOVE(arg2),arg3);
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -5229,10 +5655,10 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_CalculateEntropy__SWIG_0(PyObject *self, Py_ssize_t nobjs, PyObject **swig_obj) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   float arg3 ;
-  float *arg4 = (float *) 0 ;
+  float *arg4 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   float val3 ;
@@ -5241,6 +5667,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_CalculateEntropy__SWIG_0(PyObj
   int res4 = 0 ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if ((nobjs < 4) || (nobjs > 4)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5268,7 +5695,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_CalculateEntropy__SWIG_0(PyObj
   arg4 = reinterpret_cast< float * >(argp4);
   {
     try {
-      result = ((sentencepiece::SentencePieceProcessor const *)arg1)->CalculateEntropy(arg2,arg3,arg4);
+      result = ((sentencepiece::SentencePieceProcessor const *)arg1)->CalculateEntropy(SWIG_STD_MOVE(arg2),arg3,arg4);
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -5289,7 +5716,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_CalculateEntropy__SWIG_1(PyObject *self, Py_ssize_t nobjs, PyObject **swig_obj) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   float arg3 ;
   void *argp1 = 0 ;
@@ -5298,6 +5725,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_CalculateEntropy__SWIG_1(PyObj
   int ecode3 = 0 ;
   float result;
   
+  (void)self;
   if ((nobjs < 3) || (nobjs > 3)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5320,7 +5748,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_CalculateEntropy__SWIG_1(PyObj
   arg3 = static_cast< float >(val3);
   {
     try {
-      result = (float)((sentencepiece::SentencePieceProcessor const *)arg1)->CalculateEntropy(arg2,arg3);
+      result = (float)((sentencepiece::SentencePieceProcessor const *)arg1)->CalculateEntropy(SWIG_STD_MOVE(arg2),arg3);
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -5397,12 +5825,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_GetPieceSize(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   int result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
@@ -5428,13 +5857,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_PieceToId(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   int result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_PieceToId", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5452,7 +5882,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_PieceToId(PyObject *self, PyOb
   }
   {
     try {
-      result = (int)((sentencepiece::SentencePieceProcessor const *)arg1)->PieceToId(arg2);
+      result = (int)((sentencepiece::SentencePieceProcessor const *)arg1)->PieceToId(SWIG_STD_MOVE(arg2));
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -5468,7 +5898,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IdToPiece(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   int arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
@@ -5477,6 +5907,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IdToPiece(PyObject *self, PyOb
   PyObject *swig_obj[2] ;
   std::string *result = 0 ;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_IdToPiece", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5509,7 +5940,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_GetScore(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   int arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
@@ -5518,6 +5949,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_GetScore(PyObject *self, PyObj
   PyObject *swig_obj[2] ;
   float result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_GetScore", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5547,7 +5979,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IsUnknown(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   int arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
@@ -5556,6 +5988,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IsUnknown(PyObject *self, PyOb
   PyObject *swig_obj[2] ;
   bool result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_IsUnknown", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5585,7 +6018,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IsControl(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   int arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
@@ -5594,6 +6027,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IsControl(PyObject *self, PyOb
   PyObject *swig_obj[2] ;
   bool result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_IsControl", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5623,7 +6057,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IsUnused(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   int arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
@@ -5632,6 +6066,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IsUnused(PyObject *self, PyObj
   PyObject *swig_obj[2] ;
   bool result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_IsUnused", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5661,7 +6096,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IsByte(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   int arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
@@ -5670,6 +6105,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor_IsByte(PyObject *self, PyObjec
   PyObject *swig_obj[2] ;
   bool result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_IsByte", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5699,12 +6135,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_unk_id(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   int result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
@@ -5730,12 +6167,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_bos_id(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   int result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
@@ -5761,12 +6199,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_eos_id(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   int result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
@@ -5792,12 +6231,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_pad_id(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   int result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
@@ -5823,12 +6263,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_serialized_model_proto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
@@ -5856,13 +6297,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor_LoadFromFile(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor_LoadFromFile", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -5901,7 +6343,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsIds(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   bool arg3 ;
   int arg4 ;
@@ -5929,6 +6371,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsIds(PyObject *self, P
   PyObject *swig_obj[9] ;
   std::vector< int > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__EncodeAsIds", 9, 9, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6002,7 +6445,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsPieces(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   bool arg3 ;
   int arg4 ;
@@ -6030,6 +6473,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsPieces(PyObject *self
   PyObject *swig_obj[9] ;
   std::vector< std::string > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__EncodeAsPieces", 9, 9, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6104,7 +6548,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsSerializedProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   bool arg3 ;
   int arg4 ;
@@ -6132,6 +6576,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsSerializedProto(PyObj
   PyObject *swig_obj[9] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__EncodeAsSerializedProto", 9, 9, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6202,7 +6647,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsImmutableProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   bool arg3 ;
   int arg4 ;
@@ -6230,6 +6675,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsImmutableProto(PyObje
   PyObject *swig_obj[9] ;
   sentencepiece::ImmutableSentencePieceText result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__EncodeAsImmutableProto", 9, 9, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6298,7 +6744,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsIdsBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   int arg3 ;
   bool arg4 ;
@@ -6329,6 +6775,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsIdsBatch(PyObject *se
   PyObject *swig_obj[10] ;
   std::vector< std::vector< int > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__EncodeAsIdsBatch", 10, 10, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6429,7 +6876,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsPiecesBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   int arg3 ;
   bool arg4 ;
@@ -6460,6 +6907,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsPiecesBatch(PyObject 
   PyObject *swig_obj[10] ;
   std::vector< std::vector< std::string > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__EncodeAsPiecesBatch", 10, 10, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6561,7 +7009,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsSerializedProtoBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   int arg3 ;
   bool arg4 ;
@@ -6592,6 +7040,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsSerializedProtoBatch(
   PyObject *swig_obj[10] ;
   BytesArray result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__EncodeAsSerializedProtoBatch", 10, 10, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6688,7 +7137,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsImmutableProtoBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   int arg3 ;
   bool arg4 ;
@@ -6719,6 +7168,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__EncodeAsImmutableProtoBatch(P
   PyObject *swig_obj[10] ;
   SwigValueWrapper< std::vector< sentencepiece::ImmutableSentencePieceText > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__EncodeAsImmutableProtoBatch", 10, 10, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6816,13 +7266,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIds(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< int > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   std::string result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodeIds", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6876,13 +7327,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsBytes(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< int > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodeIdsAsBytes", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6935,13 +7387,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePieces(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   std::string result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodePieces", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -6996,13 +7449,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsSerializedProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< int > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodeIdsAsSerializedProto", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7055,13 +7509,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePiecesAsSerializedProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodePiecesAsSerializedProto", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7115,13 +7570,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsImmutableProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< int > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::ImmutableSentencePieceText result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodeIdsAsImmutableProto", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7172,13 +7628,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePiecesAsImmutableProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::ImmutableSentencePieceText result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodePiecesAsImmutableProto", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7230,7 +7687,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< std::vector< int > > *arg2 = 0 ;
   int arg3 ;
   void *argp1 = 0 ;
@@ -7240,6 +7697,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsBatch(PyObject *self
   PyObject *swig_obj[3] ;
   std::vector< std::string > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodeIdsBatch", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7311,7 +7769,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsBytesBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< std::vector< int > > *arg2 = 0 ;
   int arg3 ;
   void *argp1 = 0 ;
@@ -7321,6 +7779,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsBytesBatch(PyObjec
   PyObject *swig_obj[3] ;
   BytesArray result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodeIdsAsBytesBatch", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7391,7 +7850,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsSerializedProtoBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< std::vector< int > > *arg2 = 0 ;
   int arg3 ;
   void *argp1 = 0 ;
@@ -7401,6 +7860,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsSerializedProtoBat
   PyObject *swig_obj[3] ;
   BytesArray result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodeIdsAsSerializedProtoBatch", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7471,7 +7931,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsImmutableProtoBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< std::vector< int > > *arg2 = 0 ;
   int arg3 ;
   void *argp1 = 0 ;
@@ -7481,6 +7941,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodeIdsAsImmutableProtoBatc
   PyObject *swig_obj[3] ;
   SwigValueWrapper< std::vector< sentencepiece::ImmutableSentencePieceText > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodeIdsAsImmutableProtoBatch", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7552,7 +8013,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePiecesBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< std::vector< absl::string_view > > *arg2 = 0 ;
   int arg3 ;
   void *argp1 = 0 ;
@@ -7562,6 +8023,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePiecesBatch(PyObject *s
   PyObject *swig_obj[3] ;
   std::vector< std::string > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodePiecesBatch", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7628,7 +8090,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePiecesAsSerializedProtoBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< std::vector< absl::string_view > > *arg2 = 0 ;
   int arg3 ;
   void *argp1 = 0 ;
@@ -7638,6 +8100,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePiecesAsSerializedProto
   PyObject *swig_obj[3] ;
   BytesArray result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodePiecesAsSerializedProtoBatch", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7703,7 +8166,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePiecesAsImmutableProtoBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< std::vector< absl::string_view > > *arg2 = 0 ;
   int arg3 ;
   void *argp1 = 0 ;
@@ -7713,6 +8176,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__DecodePiecesAsImmutableProtoB
   PyObject *swig_obj[3] ;
   SwigValueWrapper< std::vector< sentencepiece::ImmutableSentencePieceText > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__DecodePiecesAsImmutableProtoBatch", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7779,7 +8243,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NBestEncodeAsIds(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   bool arg4 ;
@@ -7801,6 +8265,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NBestEncodeAsIds(PyObject *se
   PyObject *swig_obj[7] ;
   std::vector< std::vector< int > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__NBestEncodeAsIds", 7, 7, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7868,7 +8333,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NBestEncodeAsPieces(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   bool arg4 ;
@@ -7890,6 +8355,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NBestEncodeAsPieces(PyObject 
   PyObject *swig_obj[7] ;
   std::vector< std::vector< std::string > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__NBestEncodeAsPieces", 7, 7, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -7958,7 +8424,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NBestEncodeAsSerializedProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   bool arg4 ;
@@ -7980,6 +8446,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NBestEncodeAsSerializedProto(
   PyObject *swig_obj[7] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__NBestEncodeAsSerializedProto", 7, 7, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8040,7 +8507,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NBestEncodeAsImmutableProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   bool arg4 ;
@@ -8062,6 +8529,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NBestEncodeAsImmutableProto(P
   PyObject *swig_obj[7] ;
   sentencepiece::ImmutableNBestSentencePieceText result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__NBestEncodeAsImmutableProto", 7, 7, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8120,7 +8588,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__SampleEncodeAndScoreAsIds(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   float arg4 ;
@@ -8151,6 +8619,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__SampleEncodeAndScoreAsIds(PyO
   PyObject *swig_obj[10] ;
   std::vector< std::pair< std::vector< int >,float > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__SampleEncodeAndScoreAsIds", 10, 10, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8233,7 +8702,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__SampleEncodeAndScoreAsPieces(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   float arg4 ;
@@ -8264,6 +8733,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__SampleEncodeAndScoreAsPieces(
   PyObject *swig_obj[10] ;
   std::vector< std::pair< std::vector< std::string >,float > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__SampleEncodeAndScoreAsPieces", 10, 10, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8347,7 +8817,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__SampleEncodeAndScoreAsSerializedProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   float arg4 ;
@@ -8378,6 +8848,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__SampleEncodeAndScoreAsSeriali
   PyObject *swig_obj[10] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__SampleEncodeAndScoreAsSerializedProto", 10, 10, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8453,7 +8924,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__SampleEncodeAndScoreAsImmutableProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   int arg3 ;
   float arg4 ;
@@ -8484,6 +8955,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__SampleEncodeAndScoreAsImmutab
   PyObject *swig_obj[10] ;
   sentencepiece::ImmutableNBestSentencePieceText result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__SampleEncodeAndScoreAsImmutableProto", 10, 10, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8557,13 +9029,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__Normalize(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   std::string result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__Normalize", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8600,13 +9073,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__NormalizeWithOffsets(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   std::pair< std::string,std::vector< size_t > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__NormalizeWithOffsets", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8650,7 +9124,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__CalculateEntropy(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   absl::string_view arg2 ;
   float arg3 ;
   void *argp1 = 0 ;
@@ -8660,6 +9134,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__CalculateEntropy(PyObject *se
   PyObject *swig_obj[3] ;
   float result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__CalculateEntropy", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8698,7 +9173,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__CalculateEntropyBatch(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::vector< absl::string_view > *arg2 = 0 ;
   float arg3 ;
   int arg4 ;
@@ -8711,6 +9186,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceProcessor__CalculateEntropyBatch(PyObjec
   PyObject *swig_obj[4] ;
   std::vector< float > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__CalculateEntropyBatch", 4, 4, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8777,13 +9253,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceProcessor__OverrideNormalizerSpec(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceProcessor *arg1 = (sentencepiece::SentencePieceProcessor *) 0 ;
+  sentencepiece::SentencePieceProcessor *arg1 = 0 ;
   std::unordered_map< std::string,std::string > *arg2 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceProcessor__OverrideNormalizerSpec", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceProcessor, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -8842,7 +9319,7 @@ fail:
 
 
 SWIGINTERN PyObject *SentencePieceProcessor_swigregister(PyObject *SWIGUNUSEDPARM(self), PyObject *args) {
-  PyObject *obj;
+  PyObject *obj = NULL;
   if (!SWIG_Python_UnpackTuple(args, "swigregister", 1, 1, &obj)) return NULL;
   SWIG_TypeNewClientData(SWIGTYPE_p_sentencepiece__SentencePieceProcessor, SWIG_NewClientData(obj));
   return SWIG_Py_Void();
@@ -8859,6 +9336,7 @@ SWIGINTERN PyObject *_wrap_SetRandomGeneratorSeed(PyObject *self, PyObject *args
   int ecode1 = 0 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   ecode1 = SWIG_AsVal_unsigned_SS_int(swig_obj[0], &val1);
@@ -8889,6 +9367,7 @@ SWIGINTERN PyObject *_wrap_SetMinLogLevel(PyObject *self, PyObject *args) {
   int ecode1 = 0 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   ecode1 = SWIG_AsVal_int(swig_obj[0], &val1);
@@ -8917,6 +9396,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceTrainer__TrainFromString(PyObject *self,
   absl::string_view arg1 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   {
@@ -8949,6 +9429,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceTrainer__TrainFromMap(PyObject *self, Py
   std::unordered_map< std::string,std::string > *arg1 = 0 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   {
@@ -9000,9 +9481,10 @@ fail:
 SWIGINTERN PyObject *_wrap_SentencePieceTrainer__TrainFromMap2(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
   std::unordered_map< std::string,std::string > *arg1 = 0 ;
-  sentencepiece::SentenceIterator *arg2 = (sentencepiece::SentenceIterator *) 0 ;
+  sentencepiece::SentenceIterator *arg2 = 0 ;
   PyObject *swig_obj[2] ;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceTrainer__TrainFromMap2", 2, 2, swig_obj)) SWIG_fail;
   {
     std::unordered_map<std::string, std::string> *out = nullptr;
@@ -9072,6 +9554,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceTrainer__TrainFromMap3(PyObject *self, P
   PyObject *swig_obj[1] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   {
@@ -9125,10 +9608,11 @@ fail:
 SWIGINTERN PyObject *_wrap_SentencePieceTrainer__TrainFromMap4(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
   std::unordered_map< std::string,std::string > *arg1 = 0 ;
-  sentencepiece::SentenceIterator *arg2 = (sentencepiece::SentenceIterator *) 0 ;
+  sentencepiece::SentenceIterator *arg2 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::bytes result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceTrainer__TrainFromMap4", 2, 2, swig_obj)) SWIG_fail;
   {
     std::unordered_map<std::string, std::string> *out = nullptr;
@@ -9195,7 +9679,7 @@ fail:
 
 
 SWIGINTERN PyObject *SentencePieceTrainer_swigregister(PyObject *SWIGUNUSEDPARM(self), PyObject *args) {
-  PyObject *obj;
+  PyObject *obj = NULL;
   if (!SWIG_Python_UnpackTuple(args, "swigregister", 1, 1, &obj)) return NULL;
   SWIG_TypeNewClientData(SWIGTYPE_p_sentencepiece__SentencePieceTrainer, SWIG_NewClientData(obj));
   return SWIG_Py_Void();
@@ -9205,6 +9689,7 @@ SWIGINTERN PyObject *_wrap_new_SentencePieceNormalizer(PyObject *self, PyObject 
   PyObject *resultobj = 0;
   sentencepiece::SentencePieceNormalizer *result = 0 ;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "new_SentencePieceNormalizer", 0, 0, 0)) SWIG_fail;
   {
     try {
@@ -9224,11 +9709,12 @@ fail:
 
 SWIGINTERN PyObject *_wrap_delete_SentencePieceNormalizer(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, SWIG_POINTER_DISOWN |  0 );
@@ -9254,13 +9740,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceNormalizer_LoadFromSerializedProto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceNormalizer_LoadFromSerializedProto", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -9278,7 +9765,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceNormalizer_LoadFromSerializedProto(PyObj
   }
   {
     try {
-      result = (arg1)->LoadFromSerializedProto(arg2);
+      result = (arg1)->LoadFromSerializedProto(SWIG_STD_MOVE(arg2));
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -9299,13 +9786,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceNormalizer_LoadFromRuleTSV(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceNormalizer_LoadFromRuleTSV", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -9323,7 +9811,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceNormalizer_LoadFromRuleTSV(PyObject *sel
   }
   {
     try {
-      result = (arg1)->LoadFromRuleTSV(arg2);
+      result = (arg1)->LoadFromRuleTSV(SWIG_STD_MOVE(arg2));
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -9344,13 +9832,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceNormalizer_LoadFromRuleName(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceNormalizer_LoadFromRuleName", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -9368,7 +9857,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceNormalizer_LoadFromRuleName(PyObject *se
   }
   {
     try {
-      result = (arg1)->LoadFromRuleName(arg2);
+      result = (arg1)->LoadFromRuleName(SWIG_STD_MOVE(arg2));
       ReleaseResultObject(resultobj);
     }
     catch (const sentencepiece::util::Status &status) {
@@ -9389,12 +9878,13 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceNormalizer_serialized_model_proto(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[1] ;
   std::string result;
   
+  (void)self;
   if (!args) SWIG_fail;
   swig_obj[0] = args;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, 0 |  0 );
@@ -9423,13 +9913,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceNormalizer_LoadFromFile(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   sentencepiece::util::Status result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceNormalizer_LoadFromFile", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -9468,13 +9959,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceNormalizer__Normalize(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   std::string result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceNormalizer__Normalize", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -9511,13 +10003,14 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceNormalizer__NormalizeWithOffsets(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   absl::string_view arg2 ;
   void *argp1 = 0 ;
   int res1 = 0 ;
   PyObject *swig_obj[2] ;
   std::pair< std::string,std::vector< size_t > > result;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceNormalizer__NormalizeWithOffsets", 2, 2, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -9561,7 +10054,7 @@ fail:
 
 SWIGINTERN PyObject *_wrap_SentencePieceNormalizer__SetProtoField(PyObject *self, PyObject *args) {
   PyObject *resultobj = 0;
-  sentencepiece::SentencePieceNormalizer *arg1 = (sentencepiece::SentencePieceNormalizer *) 0 ;
+  sentencepiece::SentencePieceNormalizer *arg1 = 0 ;
   absl::string_view arg2 ;
   bool arg3 ;
   void *argp1 = 0 ;
@@ -9570,6 +10063,7 @@ SWIGINTERN PyObject *_wrap_SentencePieceNormalizer__SetProtoField(PyObject *self
   int ecode3 = 0 ;
   PyObject *swig_obj[3] ;
   
+  (void)self;
   if (!SWIG_Python_UnpackTuple(args, "SentencePieceNormalizer__SetProtoField", 3, 3, swig_obj)) SWIG_fail;
   res1 = SWIG_ConvertPtr(swig_obj[0], &argp1,SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, 0 |  0 );
   if (!SWIG_IsOK(res1)) {
@@ -9607,7 +10101,7 @@ fail:
 
 
 SWIGINTERN PyObject *SentencePieceNormalizer_swigregister(PyObject *SWIGUNUSEDPARM(self), PyObject *args) {
-  PyObject *obj;
+  PyObject *obj = NULL;
   if (!SWIG_Python_UnpackTuple(args, "swigregister", 1, 1, &obj)) return NULL;
   SWIG_TypeNewClientData(SWIGTYPE_p_sentencepiece__SentencePieceNormalizer, SWIG_NewClientData(obj));
   return SWIG_Py_Void();
@@ -9826,15 +10320,20 @@ static swig_const_info swig_const_table[] = {
  * array. We just loop through that array, and handle each type individually.
  * First we lookup if this type has been already loaded, and if so, use the
  * loaded structure instead of the generated one. Then we have to fill in the
- * cast linked list. The cast data is initially stored in something like a
+ * cast dependencies. The cast data is initially stored in something like a
  * two-dimensional array. Each row corresponds to a type (there are the same
  * number of rows as there are in the swig_type_initial array). Each entry in
  * a column is one of the swig_cast_info structures for that type.
  * The cast_initial array is actually an array of arrays, because each row has
- * a variable number of columns. So to actually build the cast linked list,
- * we find the array of casts associated with the type, and loop through it
- * adding the casts to the list. The one last trick we need to do is making
- * sure the type pointer in the swig_cast_info struct is correct.
+ * a variable number of columns.
+ *
+ * We loop through the cast[] array associated with the type and mark casts
+ * which have not been defined in previously loaded modules by assigning
+ * cast pointer value to cast->next. We also hash cast->type->name string
+ * and store the value in the cast->value field. If we encounter swig_cast_info
+ * structure that represents a cast to self we move it to the beginning
+ * of the cast array. One trick we need to do is making sure the type pointer
+ * in the swig_cast_info struct is correct.
  *
  * First off, we lookup the cast->type name to see if it is already loaded.
  * There are three cases to handle:
@@ -9846,8 +10345,71 @@ static swig_const_info swig_const_table[] = {
  *     cast->type) are loaded, THEN the cast info has already been loaded by
  *     the previous module so we just ignore it.
  *  3) Finally, if cast->type has not already been loaded, then we add that
- *     swig_cast_info to the linked list (because the cast->type) pointer will
+ *     swig_cast_info to the list (because the cast->type) pointer will
  *     be correct.
+ *
+ * Once the cast array has been set up AND it does have new casts that need
+ * to be added we sort non-self cast entries to move filtered out entries
+ * to the end of the array and to arrange the rest in the increasing order
+ * of their type pointer values. We store the index of the last added entry
+ * in the cast->value field of the entry[0] (overwriting the name hash). Then
+ * we sort <next|value> fields of the remaining entries to arrange hash values
+ * in the increasing order. This way cast->next->type->name field matches
+ * the cast->value hash.
+ *
+ * Example:
+ * Array of casts for type stored at 0x5000, cast to type stored at 0x3000
+ * has already been loaded
+ *
+ *      After sweep-and-hash:   After sort-by-type:    After sort-by-hash:
+ *        ________________       ________________       ________________
+ *       |                |     |                |     |                |
+ * Entry |  type = 0x5000 |     |  type = 0x5000 |     |  type = 0x5000 |
+ *   0   |                |     |                |     |                |
+ *       |  next = Entry0 |     |  next = Entry0 |     |  next = Entry0 |
+ *       | value = 1212   |     | value = 3      |     | value = 3      |
+ *       |                |     |                |     |                |
+ *       |================|     |================|     |================|
+ *       |                |     |                |     |                |
+ * Entry |  type = 0x2000 |     |  type = 0x1000 |     |  type = 0x1000 |
+ *   1   |                |     |                |     |                |
+ *       |  next = Entry1 |     |  next = Entry1 |     |  next = Entry3 |
+ *       | value = 3434   |     | value = 4545   |     | value = 2323   |
+ *       |________________|     |________________|     |________________|
+ *       |                |     |                |     |                |
+ * Entry |  type = 0x3000 |     |  type = 0x2000 |     |  type = 0x2000 |
+ *   2   |                |     |                |     |                |
+ *       |  next = 0      |     |  next = Entry2 |     |  next = Entry2 |
+ *       | value = 0      |     | value = 3434   |     | value = 3434   |
+ *       |________________|     |________________|     |________________|
+ *       |                |     |                |     |                |
+ * Entry |  type = 0x1000 |     |  type = 0x4000 |     |  type = 0x4000 |
+ *   3   |                |     |                |     |                |
+ *       |  next = Entry3 |     |  next = Entry3 |     |  next = Entry1 |
+ *       | value = 4545   |     | value = 2323   |     | value = 4545   |
+ *       |________________|     |________________|     |________________|
+ *       |                |     |                |     |                |
+ * Entry |  type = 0x4000 |     |  type = 0x3000 |     |  type = 0x3000 |
+ *   4   |                |     |                |     |                |
+ *       |  next = Entry4 |     |  next = 0      |     |  next = 0      |
+ *       | value = 2323   |     | value = 0      |     | value = 0      |
+ *       |________________|     |________________|     |________________|
+ *
+ * Once the cast array has been initialized, we use cast[0]->next field to link
+ * it into the list of cast arrays for the type.
+ *              ____    ____    ____
+ * type->cast->|next|->|next|->|next|->0
+ *             |----|  |----|  |----|
+ *             |----|  |----|  |----|
+ *             |----|  |----|  |----|
+ *
+ * Subsequent cast resolution works as follows:
+ *
+ * 1. Check whether the type matches the first entry in the current cast array.
+ * 2. If not, then do a binary search over the (0:cast->value] interval using
+ *    either type address or the hash value of the type name.
+ * 3. If not found, then move over to the next cast array (cast[0]->next).
+ *
  * ----------------------------------------------------------------------------- */
 
 #ifdef __cplusplus
@@ -9864,6 +10426,46 @@ extern "C" {
 #ifndef SWIG_INIT_CLIENT_DATA_TYPE
 #define SWIG_INIT_CLIENT_DATA_TYPE void *
 #endif
+
+/*
+ * Sort function that puts cast entries with nonzero 'next' at the front
+ * of the array while ordering them by addresses of their 'type' structs.
+ */
+SWIGINTERN int SWIG_CastCmpStruct(const void *pa, const void *pb) {
+  swig_cast_info *pca = (swig_cast_info *)pa;
+  swig_cast_info *pcb = (swig_cast_info *)pb;
+  if (pca->type < pcb->type)
+  return (pca->next || pcb->next == 0) ? -1 : 1;
+  if (pca->type > pcb->type)
+  return (pcb->next || pca->next == 0) ? 1 : -1;
+  return 0;
+}
+
+/*
+ * Shell-sort 'next' and 'value' field pairs to order them by 'value'.
+ */
+SWIGINTERN void SWIG_CastHashSort(swig_cast_info *cast, int size) {
+  const int hmax = size/9;
+  int h, i;
+  for(h = 1; h <= hmax; h = 3*h+1);
+  for(; h > 0; h /= 3)
+  {
+    for(i = h; i < size; ++i)
+    {
+      swig_cast_info *p = cast[i].next;
+      unsigned int hash = cast[i].value;
+      int j = i;
+      while(j >= h && hash < cast[j-h].value)
+      {
+        cast[j].next = cast[j-h].next;
+        cast[j].value = cast[j-h].value;
+        j -= h;
+      }
+      cast[j].next = p;
+      cast[j].value = hash;
+    }
+  }
+}
 
 SWIGRUNTIME void
 SWIG_InitializeModule(SWIG_INIT_CLIENT_DATA_TYPE clientdata) {
@@ -9916,8 +10518,9 @@ SWIG_InitializeModule(SWIG_INIT_CLIENT_DATA_TYPE clientdata) {
 #endif
   for (i = 0; i < swig_module.size; ++i) {
     swig_type_info *type = 0;
-    swig_type_info *ret;
-    swig_cast_info *cast;
+    swig_type_info *target_type;
+    swig_cast_info *cast, *first;
+    int num_mapped = 0;
     
 #ifdef SWIGRUNTIME_DEBUG
     printf("SWIG_InitializeModule: type %lu %s\n", (unsigned long)i, swig_module.type_initial[i]->name);
@@ -9943,48 +10546,101 @@ SWIG_InitializeModule(SWIG_INIT_CLIENT_DATA_TYPE clientdata) {
     }
     
     /* Insert casting types */
-    cast = swig_module.cast_initial[i];
+    cast = first = swig_module.cast_initial[i];
     while (cast->type) {
       /* Don't need to add information already in the list */
-      ret = 0;
+      target_type = 0;
 #ifdef SWIGRUNTIME_DEBUG
       printf("SWIG_InitializeModule: look cast %s\n", cast->type->name);
 #endif
       if (swig_module.next != &swig_module) {
-        ret = SWIG_MangledTypeQueryModule(swig_module.next, &swig_module, cast->type->name);
+        target_type = SWIG_MangledTypeQueryModule(swig_module.next, &swig_module, cast->type->name);
+        if (target_type) {
+          /* Target type already defined in another module */
 #ifdef SWIGRUNTIME_DEBUG
-        if (ret) printf("SWIG_InitializeModule: found cast %s\n", ret->name);
+          printf("SWIG_InitializeModule: found cast %s\n", target_type->name);
 #endif
-      }
-      if (ret) {
-        if (type == swig_module.type_initial[i]) {
+          if (type == swig_module.type_initial[i]) {
 #ifdef SWIGRUNTIME_DEBUG
-          printf("SWIG_InitializeModule: skip old type %s\n", ret->name);
+            printf("SWIG_InitializeModule: skip old type %s\n", target_type->name);
 #endif
-          cast->type = ret;
-          ret = 0;
-        } else {
-          /* Check for casting already in the list */
-          swig_cast_info *ocast = SWIG_TypeCheck(ret->name, type);
+            cast->type = target_type;
+            target_type = 0;
+          } else {
+            /* Check if this cast is already in the list */
+            swig_cast_info *ocast = SWIG_TypeCheck(target_type->name, type);
 #ifdef SWIGRUNTIME_DEBUG
-          if (ocast) printf("SWIG_InitializeModule: skip old cast %s\n", ret->name);
+            if (ocast) printf("SWIG_InitializeModule: skip old cast %s\n", target_type->name);
 #endif
-          if (!ocast) ret = 0;
+            if (!ocast) target_type = 0;
+          }
         }
       }
       
-      if (!ret) {
+      if (!target_type) {
 #ifdef SWIGRUNTIME_DEBUG
         printf("SWIG_InitializeModule: adding cast %s\n", cast->type->name);
 #endif
-        if (type->cast) {
-          type->cast->prev = cast;
-          cast->next = type->cast;
+        /* Set inclusion mark for sorting */
+        cast->next = cast;
+        num_mapped++;
+        
+        if (type == cast->type) {
+#ifdef SWIGRUNTIME_DEBUG
+          printf("%s : self cast at pos [%li]\n", type->name, cast - first);
+#endif
+          if (cast - first) {
+            /* Move cast to itself to the first entry in the array */
+            
+            swig_cast_info tmp = *cast;
+            *cast = *first;
+            *first = tmp;
+          }
+          first++;
+          
+        } else {
+          cast->value = SWIG_Hash(cast->type->name, (unsigned int)strlen(cast->type->name));
         }
-        type->cast = cast;
       }
       cast++;
     }
+    
+    if (num_mapped) {
+      if (cast - first) {
+        swig_cast_info *tmp;
+        
+        /* Sort casts by type address for binary search in SWIG_TypeCheckStruct */
+        qsort(first, cast - first, sizeof(swig_cast_info), SWIG_CastCmpStruct);
+        
+        /* Remap back links for added entries */
+        cast = swig_module.cast_initial[i] + num_mapped;
+        for (tmp = first; tmp < cast; tmp++) {
+          tmp->next = tmp;
+        }
+      }
+      
+      /* Set the value field of the first entry to the index of the last added entry */
+      cast = swig_module.cast_initial[i];
+      cast->value = num_mapped - 1;
+      
+      num_mapped -= (int)(first - cast);
+      if (num_mapped > 1) {
+        /* Sort <'next','value'> pairs by 'value' for binary search in SWIG_TypeCheck */
+        
+        SWIG_CastHashSort(first, num_mapped);
+      }
+      
+      first = type->cast;
+      if (first) {
+        /* Link the current set into the list of cast arrays */
+        cast->next = first->next;
+        first->next = cast;
+      } else {
+        cast->next = 0;
+        type->cast = cast;
+      }
+    }
+    
     /* Set entry in modules->types array equal to the type */
     swig_module.types[i] = type;
   }
@@ -10015,7 +10671,6 @@ SWIG_InitializeModule(SWIG_INIT_CLIENT_DATA_TYPE clientdata) {
 SWIGRUNTIME void
 SWIG_PropagateClientData(void) {
   size_t i;
-  swig_cast_info *equiv;
   static int init_run = 0;
   
   if (init_run) return;
@@ -10023,13 +10678,16 @@ SWIG_PropagateClientData(void) {
   
   for (i = 0; i < swig_module.size; i++) {
     if (swig_module.types[i]->clientdata) {
-      equiv = swig_module.types[i]->cast;
-      while (equiv) {
-        if (!equiv->converter) {
-          if (equiv->type && !equiv->type->clientdata)
-          SWIG_TypeClientData(equiv->type, swig_module.types[i]->clientdata);
+      swig_cast_info *head, *cast;
+      head = swig_module.types[i]->cast;
+      while (head) {
+        for (cast = head; (cast - head) <= head->value; cast++) {
+          if (!cast->converter) {
+            if (cast->type && !cast->type->clientdata)
+            SWIG_TypeClientData(cast->type, swig_module.types[i]->clientdata);
+          }
         }
-        equiv = equiv->next;
+        head = head->next;
       }
     }
   }
@@ -10072,7 +10730,7 @@ extern "C" {
       }
       if (obj) {
         PyDict_SetItemString(d, constants[i].name, obj);
-        Py_DECREF(obj);
+        SWIG_Py_DECREF(obj);
       }
     }
   }
@@ -10133,6 +10791,8 @@ extern "C" {
  *  Partial Init method
  * -----------------------------------------------------------------------------*/
 
+SWIGINTERN int SWIG_mod_exec(PyObject *module);
+
 #ifdef __cplusplus
 extern "C"
 #endif
@@ -10144,21 +10804,39 @@ PyObject*
 void
 #endif
 SWIG_init(void) {
-  PyObject *m, *d, *md, *globals;
-  
 #if PY_VERSION_HEX >= 0x03000000
+  static PyModuleDef_Slot SwigSlots[] = {
+    {
+      Py_mod_exec, (void *)SWIG_mod_exec 
+    },
+    {
+      0, NULL 
+    }
+  };
+  
   static struct PyModuleDef SWIG_module = {
     PyModuleDef_HEAD_INIT,
     SWIG_name,
     NULL,
-    -1,
+    0,
     SwigMethods,
-    NULL,
+    SwigSlots,
     NULL,
     NULL,
     NULL
   };
+  
+  return PyModuleDef_Init(&SWIG_module);
+#else
+  PyObject *m = Py_InitModule(SWIG_name, SwigMethods);
+  if (m && SWIG_mod_exec(m) != 0) {
+    Py_DECREF(m);
+  }
 #endif
+}
+
+SWIGINTERN int SWIG_mod_exec(PyObject *m) {
+  PyObject *d, *md, *globals;
   
 #if defined(SWIGPYTHON_BUILTIN)
   static SwigPyClientData SwigPyObject_clientdata = {
@@ -10196,28 +10874,30 @@ SWIG_init(void) {
   (void)self;
   
   /* Metaclass is used to implement static member variables */
-  metatype = SwigPyObjectType();
+  metatype = SwigPyObjectType_Type();
   assert(metatype);
+  
+  SwigPyStaticVar_Type();
 #endif
   
   (void)globals;
   
   /* Create singletons now to avoid potential deadlocks with multi-threaded usage after module initialization */
+  SWIG_runtime_data_module();
   SWIG_This();
   SWIG_Python_TypeCache();
-  SwigPyPacked_type();
+  SwigPyPacked_Type();
+  SwigVarLink_Type();
 #ifndef SWIGPYTHON_BUILTIN
-  SwigPyObject_type();
+  SwigPyObject_Type();
 #endif
   
   /* Fix SwigMethods to carry the callback ptrs when needed */
   SWIG_Python_FixMethods(SwigMethods, swig_const_table, swig_types, swig_type_initial);
   
-#if PY_VERSION_HEX >= 0x03000000
-  m = PyModule_Create(&SWIG_module);
-#else
-  m = Py_InitModule(SWIG_name, SwigMethods);
-#endif
+#ifdef Py_GIL_DISABLED
+  PyUnstable_Module_SetGIL(m, Py_MOD_GIL_NOT_USED);
+#endif	
   
   md = d = PyModule_GetDict(m);
   (void)md;
@@ -10235,19 +10915,15 @@ SWIG_init(void) {
     SwigPyObject_clientdata.pytype = swigpyobject;
   } else if (swigpyobject->tp_basicsize != cd->pytype->tp_basicsize) {
     PyErr_SetString(PyExc_RuntimeError, "Import error: attempted to load two incompatible swig-generated modules.");
-# if PY_VERSION_HEX >= 0x03000000
-    return NULL;
-# else
-    return;
-# endif
+    return -1;
   }
   
   /* All objects have a 'this' attribute */
-  this_descr = PyDescr_NewGetSet(SwigPyObject_type(), &this_getset_def);
+  this_descr = PyDescr_NewGetSet(SwigPyObject_Type(), &this_getset_def);
   (void)this_descr;
   
   /* All objects have a 'thisown' attribute */
-  thisown_descr = PyDescr_NewGetSet(SwigPyObject_type(), &thisown_getset_def);
+  thisown_descr = PyDescr_NewGetSet(SwigPyObject_Type(), &thisown_getset_def);
   (void)thisown_descr;
   
   public_interface = PyList_New(0);
@@ -10255,7 +10931,7 @@ SWIG_init(void) {
   (void)public_symbol;
   
   PyDict_SetItemString(md, "__all__", public_interface);
-  Py_DECREF(public_interface);
+  SWIG_Py_DECREF(public_interface);
   for (i = 0; SwigMethods[i].ml_name != NULL; ++i)
   SwigPyBuiltin_AddPublicSymbol(public_interface, SwigMethods[i].ml_name);
   for (i = 0; swig_const_table[i].name != 0; ++i)
@@ -10264,10 +10940,6 @@ SWIG_init(void) {
   
   SWIG_InstallConstants(d,swig_const_table);
   
-#if PY_VERSION_HEX >= 0x03000000
-  return m;
-#else
-  return;
-#endif
+  return 0;
 }
